refactor(screenspace): split god-file into facade + 9 modules
screenspace.py (4,777 lines) becomes a thin re-export facade (~137 lines)
over nine cohesive sibling modules, with no behavior change:

- screenspace_primitives.py  pure cv2/numpy region + image-analysis primitives
- screenspace_ocr.py         cached EasyOCR readers, number/text scoring helpers
- screenspace_frames.py      ffmpeg-pipe frame extraction + ffprobe metadata
- screenspace_scans.py       the eleven per-tool scan workflows
- screenspace_heatmap.py     template/flow heatmap PNG + GIF generation
- screenspace_tools.py       AnalysisTool registry + per-frame dispatch
- screenspace_multitool.py   multitool chaining + offset joining
- screenspace_manifest.py    task/manifest persistence + event generation
- screenspace_worker.py      the background task-queue worker

The module graph is acyclic (deepest-first: primitives/heatmap -> ocr ->
frames -> scans -> tools -> multitool/manifest -> worker -> facade). The one
real cycle (MultitoolTool.scan -> scan_multitool, scan_multitool ->
check_frame_for_tool) is broken with a single function-local import inside
MultitoolTool.scan. _merge_timestamp_spans moved to primitives (shared by
scans and multitool) to keep the DAG acyclic. All comments/docstrings move
verbatim with their code; the facade keeps the 11-tool overview docstring.

The facade re-exports every public name plus the test-touched private names,
so cli.py / screenspace_server.py / screenspace_preview.py / viewer.py and
nearly all tests keep using `import screenspace; screenspace.NAME` unchanged.

Tests: the monkeypatch-through-facade pattern (patch screenspace.X, call code
that reads X from its own module namespace) can no longer work transparently
once X lives in a sibling module, so the affected patches are repointed to the
module that reads the name (screenspace_tools/_scans/_multitool/_ocr).
test_static_skip_uses_config now reads the primitives + scans sources.
