+++ screenspace.py
@@ -18,4760 +18,120 @@ Workflow: user draws regions on a frame → enqueues tasks → ScreenspaceWorker
 a background thread → results are timestamps or artifact files → state persisted to
 screenspace_manifest.json. Region coordinates are normalized (0–1); source_width/source_height
 are stored for denormalization to target video resolution.
+
+This module is a thin re-export facade. The implementation lives in cohesive sibling
+modules (imported deepest-first below); ``import screenspace; screenspace.NAME`` keeps
+resolving every public name — and the private names the test suite reaches for — from
+their new homes:
+
+  screenspace_primitives  – pure cv2/numpy region + image-analysis primitives
+  screenspace_ocr         – cached EasyOCR readers, number/text scoring helpers
+  screenspace_frames      – ffmpeg-pipe frame extraction + ffprobe metadata
+  screenspace_scans       – the eleven per-tool scan workflows
+  screenspace_heatmap     – template/flow heatmap PNG + GIF generation
+  screenspace_tools       – AnalysisTool registry + per-frame dispatch
+  screenspace_multitool   – multitool chaining + offset joining
+  screenspace_manifest    – task/manifest persistence + event generation
+  screenspace_worker      – the background task-queue worker
 """
 
-import bisect
-import copy
-import difflib
-import functools
-import math
-import queue
-import re
-import shutil
-import statistics
-import subprocess
-import threading
-import time
-import uuid
-from concurrent.futures import Future, ThreadPoolExecutor
-from datetime import datetime, timezone
-from pathlib import Path
-from collections.abc import Iterator
-from typing import TYPE_CHECKING, Any, Callable, ClassVar
+# ruff: noqa: F401
+# Deepest-first re-export so ``screenspace.NAME`` resolves from the new modules.
 
-import cv2
-import numpy as np
-
-if TYPE_CHECKING:
-    import imagehash
-
-import config
+# ``screenspace.utils`` is part of the public surface (a test monkeypatches
+# ``screenspace.utils.warning_print``); ``utils`` is a singleton module so the
+# patch is visible to every sibling that does ``import utils``.
 import utils
-import video
 
-
-# ---------------------------------------------------------------------------
-# Module-level caches
-# ---------------------------------------------------------------------------
-
-_ocr_readers: dict[tuple, Any] = {}
-_ocr_lock = threading.Lock()
-
-
-@functools.cache
-def _morph_kernel(size: int) -> np.ndarray:
-    """Return a shared square uint8 kernel for cv2 morphology ops.
-
-    cv2.morphologyEx treats the kernel as read-only, so callers share one cached
-    array instead of reallocating np.ones() per call/frame.
-    """
-    return np.ones((size, size), np.uint8)
-
-
-ScanCallback = Callable[[float, np.ndarray], bool | None]
-"""Per-frame callback signature for scan_video_frames and friends.
-
-Receives ``(timestamp_seconds, region_pixels)`` and may return ``False`` to stop
-iteration early. Used by all eight scan tools (color, change, similarity, text,
-numbers, flow, scene, inactivity)."""
-
-
-def _get_ocr_reader(languages: list[str]) -> Any:
-    """Return a cached EasyOCR Reader for the given language set."""
-    import easyocr
-
-    key = tuple(sorted(languages))
-    with _ocr_lock:
-        if key not in _ocr_readers:
-            _ocr_readers[key] = easyocr.Reader(list(key), verbose=False)
-        return _ocr_readers[key]
-
-
-def _preprocess_for_ocr(pixels: np.ndarray, *, min_height: int = 0) -> np.ndarray:
-    """Enhance a region crop for OCR: upscale small crops and boost local contrast.
-
-    Compressed game HUDs frequently render text below EasyOCR's comfortable size
-    and at low contrast. Crops shorter than ``min_height`` are upscaled with cubic
-    interpolation (preserving aspect ratio), then CLAHE (contrast-limited adaptive
-    histogram equalization) separates faint glyphs from the background. Opt-in per
-    task because the resize + equalization costs a few ms/frame and can introduce
-    ringing on already-clean text.
-
-    Returns a 3-channel BGR array (EasyOCR accepts grayscale, but stacking keeps
-    the downstream call identical to the raw crop).
-    """
-    if min_height <= 0:
-        min_height = config.SCREENSPACE_OCR_MIN_HEIGHT
-    h, w = pixels.shape[:2]
-    if h == 0 or w == 0:
-        return pixels
-    if h < min_height:
-        scale = min_height / float(h)
-        new_w = max(1, int(round(w * scale)))
-        pixels = cv2.resize(pixels, (new_w, min_height), interpolation=cv2.INTER_CUBIC)
-    gray = cv2.cvtColor(pixels, cv2.COLOR_BGR2GRAY) if pixels.ndim == 3 else pixels
-    clahe = cv2.createCLAHE(clipLimit=2.0, tileGridSize=(8, 8))
-    enhanced = clahe.apply(gray)
-    return cv2.cvtColor(enhanced, cv2.COLOR_GRAY2BGR)
-
-
-# Opt-in OCR confusion-collapsing for the text tool. Folds glyphs EasyOCR most
-# often swaps on compressed footage toward a canonical form before the fuzzy
-# compare. Two directions: fold toward digits (so a search for "100" matches an
-# OCR reading of "l00") or toward letters (so a search for "stop" matches "5top").
-_OCR_FOLD_TO_DIGITS = str.maketrans(
-    {"o": "0", "l": "1", "i": "1", "|": "1", "s": "5", "b": "8"}
+from screenspace_primitives import (
+    FULL_FRAME_REGION,
+    FULL_FRAME_REGION_NAME,
+    ScanCallback,
+    _ConsecutiveBuffer,
+    _is_static_skip,
+    _merge_timestamp_spans,
+    _morph_kernel,
+    _prepare_template,
+    average_color_hsv,
+    color_matches,
+    color_present,
+    compare_scene_fingerprints,
+    compute_frame_diff,
+    compute_optical_flow,
+    compute_phash,
+    compute_scene_fingerprint,
+    denormalize_region,
+    extract_region,
+    match_template,
+    regions_are_similar,
+    resolve_region_request,
 )
-_OCR_FOLD_TO_LETTERS = str.maketrans({"0": "o", "1": "l", "5": "s", "8": "b"})
-
-
-def _normalize_ocr_text(s: str, mode: str) -> str:
-    """Lowercase, then fold easily-confused glyphs toward ``mode``'s canonical form.
-
-    ``mode`` is ``"digits"`` (fold letters→digits), ``"letters"`` (fold
-    digits→letters), or ``"off"`` (lowercase only — also the fallback for any
-    unrecognized value).
-    """
-    s = s.lower()
-    if mode == "digits":
-        return s.translate(_OCR_FOLD_TO_DIGITS)
-    if mode == "letters":
-        return s.translate(_OCR_FOLD_TO_LETTERS)
-    return s
-
-
-def _effective_ocr_confidence_threshold(value: Any = None) -> float:
-    """Return OCR confidence cutoff, using config default only when omitted."""
-    if value is None:
-        return config.SCREENSPACE_OCR_MIN_CONFIDENCE
-    try:
-        threshold = float(value)
-    except (TypeError, ValueError) as exc:
-        raise ValueError("ocr_confidence_threshold must be a number") from exc
-    if not math.isfinite(threshold):
-        raise ValueError("ocr_confidence_threshold must be a finite number")
-    if threshold < 0 or threshold > 1:
-        raise ValueError("ocr_confidence_threshold must be between 0 and 1")
-    return threshold
-
-
-class _ConsecutiveBuffer:
-    """Emit one event only after N consecutive matching sampled frames.
-
-    ``push()`` records a matching frame; once ``size`` matches accumulate it
-    returns a single event -- re-stamped with the median timestamp of the run and
-    carrying the payload of the frame nearest that median -- and resets.
-    ``reset()`` (called on any non-match) discards a partial run. ``size == 1``
-    emits on every push, so the default reproduces pre-temporal-coherence
-    behavior exactly.
-    """
-
-    def __init__(self, size: int) -> None:
-        self.size = max(1, int(size))
-        self._events: list[dict[str, Any]] = []
-        self._timestamps: list[float] = []
-
-    def push(self, ts: float, event: dict[str, Any]) -> dict[str, Any] | None:
-        self._events.append(event)
-        self._timestamps.append(ts)
-        if len(self._events) >= self.size:
-            median_ts = statistics.median(self._timestamps)
-            # Pair the median timestamp with the payload of the frame closest to
-            # it. For odd-length runs this is the exact middle frame; for even
-            # lengths (the median is interpolated between two frames) it picks
-            # the nearer real frame instead of an arbitrary upper-middle one.
-            nearest = min(
-                range(len(self._timestamps)),
-                key=lambda i: abs(self._timestamps[i] - median_ts),
-            )
-            emitted = dict(self._events[nearest])
-            emitted["timestamp"] = median_ts
-            self.reset()
-            return emitted
-        return None
-
-    def carry(self, ts: float) -> dict[str, Any] | None:
-        """Extend an active run when a frame is skipped as static.
-
-        A static frame is near-identical to the last processed frame, so a match
-        that was on screen is still there -- the run should continue, not break.
-        Re-pushes the most recent matched event under *ts* (returning an emit if
-        that completes the run). No-op when no run is active (nothing to carry),
-        which keeps ``size == 1`` -- where ``push`` emits and resets every frame
-        -- on its legacy path.
-        """
-        if not self._events:
-            return None
-        return self.push(ts, self._events[-1])
-
-    def reset(self) -> None:
-        self._events = []
-        self._timestamps = []
-
-
-def _is_static_skip(
-    ts: float,
-    pixels: np.ndarray,
-    prev_gray: list[np.ndarray | None],
-    buf: _ConsecutiveBuffer,
-    results: list[dict[str, Any]],
-    on_result: Callable[[dict[str, Any]], None] | None,
-    on_progress: Callable[[float], None] | None,
-    start_seconds: float,
-    total_range: float,
-) -> bool:
-    """Decide whether *pixels* is a near-duplicate of the previous frame.
-
-    Returns True when the mean grayscale diff from the last processed frame is
-    below ``SCREENSPACE_STATIC_FRAME_SKIP_THRESHOLD`` -- the caller should then
-    ``return None`` from its per-frame callback. The content (and any active
-    match) is unchanged, so carry the consecutive-match run via ``buf.carry``
-    (emitting through *results*/*on_result*) and report progress rather than
-    breaking the run. Otherwise records *pixels* as the new baseline and returns
-    False so the caller runs its real analysis. Shared by scan_text/scan_numbers.
-    """
-    gray = cv2.cvtColor(pixels, cv2.COLOR_BGR2GRAY)
-    if prev_gray[0] is not None:
-        diff = float(np.mean(cv2.absdiff(prev_gray[0], gray)))
-        if diff < config.SCREENSPACE_STATIC_FRAME_SKIP_THRESHOLD:
-            emitted = buf.carry(ts)
-            if emitted is not None:
-                results.append(emitted)
-                if on_result:
-                    on_result(emitted)
-            if on_progress and total_range > 0:
-                on_progress((ts - start_seconds) / total_range)
-            return True
-    prev_gray[0] = gray
-    return False
-
-
-# ---------------------------------------------------------------------------
-# Analysis primitives
-# ---------------------------------------------------------------------------
-
-
-def extract_region(frame: np.ndarray, region: dict[str, int]) -> np.ndarray:
-    """Crop a rectangular region from a frame.
-
-    Args:
-        frame: BGR image as numpy array (H, W, 3).
-        region: Dict with keys ``x``, ``y``, ``w``, ``h`` in pixels.
-
-    Returns:
-        Cropped region as numpy array.
-    """
-    h_frame, w_frame = frame.shape[:2]
-    x = max(0, region["x"])
-    y = max(0, region["y"])
-    x2 = min(w_frame, x + region["w"])
-    y2 = min(h_frame, y + region["h"])
-    return frame[y:y2, x:x2]
-
-
-def denormalize_region(
-    region: dict[str, Any], target_w: int, target_h: int
-) -> dict[str, int]:
-    """Convert a normalized region (0–1 floats) to pixel coordinates.
-
-    Args:
-        region: Dict with normalized ``x``, ``y``, ``w``, ``h`` keys.
-        target_w: Target frame width in pixels.
-        target_h: Target frame height in pixels.
-
-    Returns:
-        Dict with integer pixel ``x``, ``y``, ``w``, ``h`` keys.
-    """
-    return {
-        "x": int(round(region["x"] * target_w)),
-        "y": int(round(region["y"] * target_h)),
-        "w": int(round(region["w"] * target_w)),
-        "h": int(round(region["h"] * target_h)),
-    }
-
-
-FULL_FRAME_REGION_NAME = "full_frame"
-FULL_FRAME_REGION: dict[str, Any] = {
-    "x": 0.0,
-    "y": 0.0,
-    "w": 1.0,
-    "h": 1.0,
-    "source_width": 0,
-    "source_height": 0,
-}
-
-
-def resolve_region_request(
-    region_name: str,
-    region_ref: Any,
-    manifest: dict[str, Any],
-) -> tuple[str, dict[str, Any]]:
-    """Resolve a (region_name, region_ref) request to (resolved_name, normalized_region).
-
-    Mirrors the active/stash/full_frame precedence the Screenspace server and CLI both
-    rely on: a ``region_ref`` (``{source: "active"|"stash"|"full_frame", name?, stash_id?}``)
-    pins the lookup to a specific source, while a bare ``region_name`` falls back to
-    active-regions-first then stashes. Pure — reads only the passed ``manifest`` and never
-    flattens stashes, so duplicate region names across stashes stay distinguishable.
-
-    Returns the resolved name and the normalized (0–1) region dict. Raises ``ValueError``
-    with a human-readable message on any failure (caller maps it to a 400 or a CLI error).
-    """
-    active_regions = manifest.get("regions", {})
-    if not isinstance(active_regions, dict):
-        active_regions = {}
-
-    if region_ref is None:
-        if region_name == FULL_FRAME_REGION_NAME:
-            return FULL_FRAME_REGION_NAME, dict(FULL_FRAME_REGION)
-        if region_name in active_regions:
-            return region_name, active_regions[region_name]
-        for stash in manifest.get("stashes", []):
-            stash_regions = stash.get("regions", {})
-            if isinstance(stash_regions, dict) and region_name in stash_regions:
-                return region_name, stash_regions[region_name]
-        raise ValueError(f"Region '{region_name}' not found")
-
-    if not isinstance(region_ref, dict):
-        raise ValueError("region_ref must be an object")
-
-    source = str(region_ref.get("source", "")).strip()
-
-    if source == "full_frame":
-        return FULL_FRAME_REGION_NAME, dict(FULL_FRAME_REGION)
-
-    name = str(region_ref.get("name", "")).strip()
-    if not name:
-        raise ValueError("region_ref.name is required")
-
-    if source == "active":
-        if name not in active_regions:
-            raise ValueError(f"Region '{name}' not found")
-        return name, active_regions[name]
-
-    if source == "stash":
-        stash_id = str(region_ref.get("stash_id", "")).strip()
-        if not stash_id:
-            raise ValueError("region_ref.stash_id is required")
-        for stash in manifest.get("stashes", []):
-            if stash.get("id") != stash_id:
-                continue
-            stash_regions = stash.get("regions", {})
-            if not isinstance(stash_regions, dict) or name not in stash_regions:
-                raise ValueError(f"Region '{name}' not found in stash '{stash_id}'")
-            return name, stash_regions[name]
-        raise ValueError(f"Stash '{stash_id}' not found")
-
-    raise ValueError("region_ref.source must be 'active', 'stash', or 'full_frame'")
-
-
-def average_color_hsv(region_pixels: np.ndarray) -> dict[str, float]:
-    """Compute mean HSV color of a region.
-
-    Args:
-        region_pixels: BGR image region as numpy array.
-
-    Returns:
-        Dict with keys ``h`` (0-180), ``s`` (0-255), ``v`` (0-255).
-    """
-    h, w = region_pixels.shape[:2]
-    if h > 64 or w > 64:
-        region_pixels = cv2.resize(
-            region_pixels, (min(w, 64), min(h, 64)), interpolation=cv2.INTER_AREA
-        )
-    hsv = cv2.cvtColor(region_pixels, cv2.COLOR_BGR2HSV)
-    mean = np.mean(hsv, axis=(0, 1))
-    return {"h": float(mean[0]), "s": float(mean[1]), "v": float(mean[2])}
-
-
-def color_matches(
-    region_pixels: np.ndarray,
-    target_color: dict[str, float],
-    tolerance: dict[str, float],
-) -> tuple[bool, float]:
-    """Check if region's average HSV color is within tolerance of target.
-
-    Handles hue wraparound (red at 0/180 boundary).
-
-    Returns:
-        Tuple of (matches, confidence) where confidence is 0.0–1.0.
-    """
-    avg = average_color_hsv(region_pixels)
-    hue_diff = abs(avg["h"] - target_color["h"])
-    hue_dist = min(hue_diff, 180.0 - hue_diff)
-    s_dist = abs(avg["s"] - target_color["s"])
-    v_dist = abs(avg["v"] - target_color["v"])
-    matched = (
-        hue_dist <= tolerance["h"]
-        and s_dist <= tolerance["s"]
-        and v_dist <= tolerance["v"]
-    )
-    conf = max(
-        0.0,
-        1.0
-        - max(
-            hue_dist / max(tolerance["h"], 1e-6),
-            s_dist / max(tolerance["s"], 1e-6),
-            v_dist / max(tolerance["v"], 1e-6),
-        ),
-    )
-    return matched, conf
-
-
-def color_present(
-    region_pixels: np.ndarray,
-    target_color: dict[str, float],
-    tolerance: dict[str, float],
-    min_coverage: float = 0.0,
-) -> tuple[bool, float]:
-    """Check whether the target color appears *anywhere* in the region.
-
-    Unlike :func:`color_matches` (which averages the whole region), this builds
-    a per-pixel HSV match mask and fires when the fraction of matching pixels
-    reaches ``min_coverage``. A small element of the target color in an
-    otherwise neutral region is detected here but averaged away by
-    :func:`color_matches`. Hue wraparound (red at the 0/180 boundary) is handled
-    the same way as :func:`color_matches`.
-
-    The region is scanned at full resolution (no ``INTER_AREA`` downscale) so
-    small patches survive; callers running in fast-scan mode must avoid the
-    ``max_region_dim`` downscale for this path.
-
-    Returns:
-        Tuple of (matches, coverage) where ``coverage`` is the 0.0–1.0 fraction
-        of pixels matching the target. With ``min_coverage <= 0`` a single
-        matching pixel fires; otherwise ``coverage >= min_coverage`` is required.
-    """
-    if region_pixels.size == 0:
-        return False, 0.0
-    hsv = cv2.cvtColor(region_pixels, cv2.COLOR_BGR2HSV)
-    h = hsv[..., 0].astype(np.float32)
-    s = hsv[..., 1].astype(np.float32)
-    v = hsv[..., 2].astype(np.float32)
-    hue_diff = np.abs(h - float(target_color["h"]))
-    hue_dist = np.minimum(hue_diff, 180.0 - hue_diff)
-    mask = (
-        (hue_dist <= tolerance["h"])
-        & (np.abs(s - float(target_color["s"])) <= tolerance["s"])
-        & (np.abs(v - float(target_color["v"])) <= tolerance["v"])
-    )
-    count = int(np.count_nonzero(mask))
-    coverage = count / mask.size if mask.size else 0.0
-    matched = count > 0 if min_coverage <= 0 else coverage >= min_coverage
-    return matched, float(coverage)
-
-
-def compute_frame_diff(
-    region_a: np.ndarray,
-    region_b: np.ndarray,
-    noise_threshold: int = 0,
-) -> float:
-    """Compute pixel difference ratio between two same-sized regions.
-
-    Applies Gaussian blur, thresholds noise, and morphological opening.
-
-    Returns:
-        Change ratio 0.0-1.0 (fraction of pixels that changed).
-    """
-    if noise_threshold <= 0:
-        noise_threshold = config.SCREENSPACE_NOISE_THRESHOLD
-    k = config.SCREENSPACE_BLUR_KERNEL
-    a_blur = cv2.GaussianBlur(region_a, (k, k), 0)
-    b_blur = cv2.GaussianBlur(region_b, (k, k), 0)
-    a_gray = cv2.cvtColor(a_blur, cv2.COLOR_BGR2GRAY)
-    b_gray = cv2.cvtColor(b_blur, cv2.COLOR_BGR2GRAY)
-    diff = cv2.absdiff(a_gray, b_gray)
-    _, mask = cv2.threshold(diff, noise_threshold, 255, cv2.THRESH_BINARY)
-    kernel = _morph_kernel(config.SCREENSPACE_MORPH_KERNEL)
-    mask = cv2.morphologyEx(mask, cv2.MORPH_OPEN, kernel)
-    if mask.size == 0:
-        return 0.0
-    return float(np.count_nonzero(mask)) / float(mask.size)
-
-
-def regions_are_similar(
-    region_a: np.ndarray,
-    region_b: np.ndarray,
-    threshold: float = 0.0,
-) -> tuple[bool, float]:
-    """SSIM-based similarity check with blur preprocessing.
-
-    Returns:
-        Tuple of (is_similar, ssim_score).
-    """
-    if threshold <= 0.0:
-        threshold = config.SCREENSPACE_SSIM_THRESHOLD
-    max_dim = 256
-    h, w = region_a.shape[:2]
-    if h > max_dim or w > max_dim:
-        scale = max_dim / max(h, w)
-        new_w, new_h = int(w * scale), int(h * scale)
-        region_a = cv2.resize(region_a, (new_w, new_h), interpolation=cv2.INTER_AREA)
-        region_b = cv2.resize(region_b, (new_w, new_h), interpolation=cv2.INTER_AREA)
-    k = config.SCREENSPACE_BLUR_KERNEL
-    a_blur = cv2.GaussianBlur(region_a, (k, k), 0)
-    b_blur = cv2.GaussianBlur(region_b, (k, k), 0)
-    a_gray = cv2.cvtColor(a_blur, cv2.COLOR_BGR2GRAY)
-    b_gray = cv2.cvtColor(b_blur, cv2.COLOR_BGR2GRAY)
-    from skimage.metrics import structural_similarity as ssim
-
-    score = float(ssim(a_gray, b_gray))
-    return score >= threshold, score
-
-
-def compute_phash(region_pixels: np.ndarray) -> "imagehash.ImageHash":
-    """Compute perceptual hash of a region for fast similarity scanning."""
-    import imagehash
-    from PIL import Image
-
-    rgb = cv2.cvtColor(region_pixels, cv2.COLOR_BGR2RGB)
-    pil_img = Image.fromarray(rgb)
-    return imagehash.phash(pil_img)
-
-
-# Prepared template payload shared across frames in a scan: grayscale blurred
-# template, grayscale blurred mask (or None), and a "degenerate" flag set when
-# the template has near-zero variance (TM_CCOEFF_NORMED is undefined there).
-_PreparedTemplate = tuple[np.ndarray, "np.ndarray | None", bool]
-
-
-def _scale_template(
-    template: np.ndarray,
-    mask: np.ndarray | None,
-    template_scale: float,
-) -> tuple[np.ndarray, np.ndarray | None]:
-    """Resize a template (and its mask) for matching.
-
-    Combines the user-supplied *template_scale* with the global CV resolution
-    scale so the template matches at the same relative size on the (possibly
-    upscaled) extracted frames. The opt-in template_scale slider lets users
-    compensate when their uploaded PNG is captured at a different pixel scale
-    than its in-video rendering (e.g. a 50x50 icon appearing as 24x24 on
-    screen). Returns the pair unchanged when the effective scale is ~1.0.
-    """
-    cv_scale = (
-        config.SCREENSPACE_CV_RESOLUTION_SCALE
-        if config.SCREENSPACE_CV_RESOLUTION_SCALE > 0
-        else 1.0
-    )
-    effective = template_scale * cv_scale
-    if not (effective > 0 and abs(effective - 1.0) > 1e-6):
-        return template, mask
-    th, tw = template.shape[:2]
-    nw = max(8, int(round(tw * effective)))
-    nh = max(8, int(round(th * effective)))
-    interp = cv2.INTER_AREA if effective < 1.0 else cv2.INTER_CUBIC
-    scaled_template = cv2.resize(template, (nw, nh), interpolation=interp)
-    scaled_mask = (
-        cv2.resize(mask, (nw, nh), interpolation=cv2.INTER_NEAREST)
-        if mask is not None
-        else None
-    )
-    return scaled_template, scaled_mask
-
-
-def _prepare_template(
-    template: np.ndarray, mask: np.ndarray | None
-) -> _PreparedTemplate:
-    """Compute the per-scan-constant grayscale template and mask once.
-
-    Hoisted out of :func:`match_template` so callers that run the same
-    template against many frames (scan_template, evaluate_region) can pay
-    the blur+cvtColor cost a single time instead of per-frame.
-    """
-    k = config.SCREENSPACE_BLUR_KERNEL
-    tmpl_gray = cv2.cvtColor(cv2.GaussianBlur(template, (k, k), 0), cv2.COLOR_BGR2GRAY)
-    # Binarize the alpha mask (>= 128 -> 255, else 0) instead of blurring it.
-    # Soft-blurred masks let semi-transparent edge pixels contribute partially
-    # to cv2.matchTemplate, which inflates TM_CCOEFF_NORMED scores for
-    # mostly-transparent PNG icons and produces false positives.
-    if mask is not None:
-        _, gray_mask = cv2.threshold(mask, 127, 255, cv2.THRESH_BINARY)
-    else:
-        gray_mask = None
-    # Degenerate if the pixels that will actually contribute to matching have
-    # no variance. TM_CCOEFF_NORMED normalizes by the masked template std —
-    # when that is ~0 (e.g. a mostly-transparent PNG with a flat opaque patch,
-    # especially after scaling down) the denominator underflows and every
-    # position gets a near-1.0 score.
-    if gray_mask is not None:
-        masked = tmpl_gray[gray_mask > 0]
-        contributing_std = float(masked.std()) if masked.size else 0.0
-    else:
-        contributing_std = float(np.std(tmpl_gray))
-    degenerate = contributing_std < 1.0
-    return (tmpl_gray, gray_mask, degenerate)
-
-
-def _template_correlation_map(
-    frame: np.ndarray, prepared: _PreparedTemplate
-) -> np.ndarray | None:
-    """Compute the finite TM_CCOEFF_NORMED correlation map for a prepared template.
-
-    Returns ``None`` when matching is undefined — a degenerate (constant)
-    template, or one larger than the frame. inf/nan cells (near-zero variance
-    patches) are neutralized to ``-1.0`` so callers can safely threshold the map
-    or read its peak (the threshold-independent scalar used by calibration).
-    """
-    tmpl_gray, gray_mask, degenerate = prepared
-    if degenerate:
-        # A constant (zero-variance) template produces undefined TM_CCOEFF_NORMED
-        # results — every position may score ~1.0.  Bail out.
-        return None
-    k = config.SCREENSPACE_BLUR_KERNEL
-    frame_gray = cv2.cvtColor(cv2.GaussianBlur(frame, (k, k), 0), cv2.COLOR_BGR2GRAY)
-    th, tw = tmpl_gray.shape[:2]
-    if th > frame_gray.shape[0] or tw > frame_gray.shape[1]:
-        return None
-    result = cv2.matchTemplate(
-        frame_gray, tmpl_gray, cv2.TM_CCOEFF_NORMED, mask=gray_mask
-    )
-    if not np.all(np.isfinite(result)):
-        result = np.where(np.isfinite(result), result, -1.0)
-    return result
-
-
-def _match_template_prepared(
-    frame: np.ndarray,
-    prepared: _PreparedTemplate,
-    threshold: float,
-    nms_overlap: float,
-) -> list[dict[str, Any]]:
-    """Match a frame against an already-prepared template payload."""
-    result = _template_correlation_map(frame, prepared)
-    if result is None:
-        return []
-    tmpl_gray, _gray_mask, _degenerate = prepared
-    th, tw = tmpl_gray.shape[:2]
-    locs = np.where(result >= threshold)
-    if len(locs[0]) == 0:
-        return []
-
-    # Guard against pathological matchTemplate outputs (e.g. low-variance
-    # masked templates at certain scales) that can produce tens of thousands
-    # of above-threshold candidates. The O(n^2) NMS below would otherwise
-    # freeze the worker. Cap to the top _MAX_CANDIDATES by raw score.
-    _MAX_CANDIDATES = 5000
-    scores = result[locs]
-    if len(locs[0]) > _MAX_CANDIDATES:
-        top_idx = np.argpartition(scores, -_MAX_CANDIDATES)[-_MAX_CANDIDATES:]
-        ys, xs = locs[0][top_idx], locs[1][top_idx]
-        scores = scores[top_idx]
-    else:
-        ys, xs = locs[0], locs[1]
-
-    detections: list[dict[str, Any]] = []
-    for pt_y, pt_x, raw in zip(ys, xs, scores):
-        score = float(raw)
-        if not math.isfinite(score):
-            continue
-        detections.append(
-            {"x": int(pt_x), "y": int(pt_y), "w": tw, "h": th, "score": score}
-        )
-    detections.sort(key=lambda d: d["score"], reverse=True)
-
-    # Non-maximum suppression
-    kept: list[dict[str, Any]] = []
-    for det in detections:
-        overlaps = False
-        for k_det in kept:
-            # Compute IoU
-            xa = max(det["x"], k_det["x"])
-            ya = max(det["y"], k_det["y"])
-            xb = min(det["x"] + det["w"], k_det["x"] + k_det["w"])
-            yb = min(det["y"] + det["h"], k_det["y"] + k_det["h"])
-            inter = max(0, xb - xa) * max(0, yb - ya)
-            area_a = det["w"] * det["h"]
-            area_b = k_det["w"] * k_det["h"]
-            union = area_a + area_b - inter
-            if union > 0 and inter / union > nms_overlap:
-                overlaps = True
-                break
-        if not overlaps:
-            kept.append(det)
-    return kept
-
-
-def match_template(
-    frame: np.ndarray,
-    template: np.ndarray,
-    threshold: float = 0.0,
-    nms_overlap: float = 0.0,
-    mask: np.ndarray | None = None,
-    *,
-    prepared: _PreparedTemplate | None = None,
-) -> list[dict[str, Any]]:
-    """Find all locations where template appears in frame.
-
-    Uses ``cv2.matchTemplate`` with ``TM_CCOEFF_NORMED``.  Non-maximum
-    suppression removes overlapping detections.  An optional *mask*
-    (same size as *template*, single-channel) restricts matching to
-    non-transparent regions — useful for uploaded PNGs with alpha.
-
-    When calling repeatedly with the same *template* / *mask* (e.g. one
-    scan over many frames), build a *prepared* tuple once with
-    :func:`_prepare_template` and pass it in to skip the per-call blur
-    and grayscale conversion of the template.
-
-    Returns:
-        List of ``{x, y, w, h, score}`` dicts for each match above *threshold*.
-    """
-    if threshold <= 0.0:
-        threshold = config.SCREENSPACE_TEMPLATE_MATCH_THRESHOLD
-    if nms_overlap <= 0.0:
-        nms_overlap = config.SCREENSPACE_TEMPLATE_NMS_OVERLAP
-
-    if prepared is None:
-        prepared = _prepare_template(template, mask)
-    return _match_template_prepared(frame, prepared, threshold, nms_overlap)
-
-
-def compute_optical_flow(
-    prev_gray: np.ndarray,
-    curr_gray: np.ndarray,
-    pyr_scale: float = 0.0,
-    return_grid: bool = False,
-) -> dict[str, Any]:
-    """Compute dense optical flow between two grayscale frames.
-
-    Returns:
-        Dict with ``magnitude`` (mean flow vector length),
-        ``angle`` (dominant direction in degrees, 0-360), and optionally
-        ``flow_grid`` (sparse grid of motion vectors for visualization).
-    """
-    if pyr_scale <= 0.0:
-        pyr_scale = config.SCREENSPACE_FLOW_PYR_SCALE
-
-    # Resize to max 256px for speed
-    max_dim = 256
-    h, w = prev_gray.shape[:2]
-    if h > max_dim or w > max_dim:
-        scale = max_dim / max(h, w)
-        new_w, new_h = int(w * scale), int(h * scale)
-        prev_gray = cv2.resize(prev_gray, (new_w, new_h), interpolation=cv2.INTER_AREA)
-        curr_gray = cv2.resize(curr_gray, (new_w, new_h), interpolation=cv2.INTER_AREA)
-
-    flow_out = np.zeros((*prev_gray.shape[:2], 2), dtype=np.float32)
-    flow = cv2.calcOpticalFlowFarneback(
-        prev_gray, curr_gray, flow_out, pyr_scale, 3, 15, 3, 5, 1.2, 0
-    )
-    mag, ang = cv2.cartToPolar(flow[..., 0], flow[..., 1], angleInDegrees=True)
-    mean_mag = float(np.mean(mag))
-
-    # Dominant angle: weighted mean by magnitude
-    if mean_mag > 0:
-        # Use circular mean to avoid wraparound issues
-        rad = np.deg2rad(ang)
-        sin_sum = float(np.sum(mag * np.sin(rad)))
-        cos_sum = float(np.sum(mag * np.cos(rad)))
-        dominant_angle = float(np.rad2deg(np.arctan2(sin_sum, cos_sum))) % 360.0
-    else:
-        dominant_angle = 0.0
-
-    result: dict[str, Any] = {
-        "magnitude": round(mean_mag, 4),
-        "angle": round(dominant_angle, 1),
-    }
-
-    if return_grid:
-        grid_size = config.SCREENSPACE_FLOW_GRID_SIZE
-        min_mag = config.SCREENSPACE_FLOW_GRID_MIN_MAG
-        gh, gw = mag.shape[:2]
-        step_y = max(1, gh // grid_size)
-        step_x = max(1, gw // grid_size)
-        grid: list[dict[str, float]] = []
-        for gy in range(0, gh, step_y):
-            for gx in range(0, gw, step_x):
-                cell_mag = float(np.mean(mag[gy : gy + step_y, gx : gx + step_x]))
-                if cell_mag < min_mag:
-                    continue
-                cell_ang = float(np.mean(ang[gy : gy + step_y, gx : gx + step_x]))
-                grid.append(
-                    {
-                        "x": round((gx + step_x / 2) / gw, 3),
-                        "y": round((gy + step_y / 2) / gh, 3),
-                        "mag": round(cell_mag, 2),
-                        "ang": round(cell_ang, 1),
-                    }
-                )
-        result["flow_grid"] = grid
-
-    return result
-
-
-def compute_scene_fingerprint(region_pixels: np.ndarray) -> dict[str, Any]:
-    """Compute a feature-based fingerprint for scene classification.
-
-    Combines HSV histogram, edge density, and color statistics into a
-    fingerprint suitable for comparison via :func:`compare_scene_fingerprints`.
-    """
-    # Resize to standardize
-    max_dim = 128
-    h, w = region_pixels.shape[:2]
-    if h > max_dim or w > max_dim:
-        scale = max_dim / max(h, w)
-        region_pixels = cv2.resize(
-            region_pixels,
-            (int(w * scale), int(h * scale)),
-            interpolation=cv2.INTER_AREA,
-        )
-
-    bins = config.SCREENSPACE_SCENE_HISTOGRAM_BINS
-    hsv = cv2.cvtColor(region_pixels, cv2.COLOR_BGR2HSV)
-    # 3D histogram flattened
-    hist = cv2.calcHist(
-        [hsv],
-        [0, 1, 2],
-        None,
-        [bins, bins, bins],
-        [0, 180, 0, 256, 0, 256],
-    )
-    cv2.normalize(hist, hist)
-
-    # Edge density
-    gray = cv2.cvtColor(region_pixels, cv2.COLOR_BGR2GRAY)
-    edges = cv2.Canny(gray, 100, 200)
-    edge_density = (
-        float(np.count_nonzero(edges)) / float(edges.size) if edges.size > 0 else 0.0
-    )
-
-    # Color stats per channel
-    color_stats: list[float] = []
-    for ch in range(3):
-        channel = region_pixels[:, :, ch].astype(np.float64)
-        color_stats.extend([float(np.mean(channel)), float(np.std(channel))])
-
-    return {
-        "histogram": hist,
-        "edge_density": edge_density,
-        "color_stats": color_stats,
-    }
-
-
-def compare_scene_fingerprints(
-    fp_a: dict[str, Any],
-    fp_b: dict[str, Any],
-) -> float:
-    """Compare two scene fingerprints.
-
-    Returns similarity score 0.0–1.0.
-    """
-    # Histogram correlation: range [-1, 1] → [0, 1]
-    # Flatten 3D histograms to 1D — cv2.compareHist returns incorrect
-    # results for multidimensional arrays.
-    hist_corr = cv2.compareHist(
-        fp_a["histogram"].flatten().astype(np.float32),
-        fp_b["histogram"].flatten().astype(np.float32),
-        cv2.HISTCMP_CORREL,
-    )
-    hist_sim = (hist_corr + 1.0) / 2.0
-
-    # Edge density similarity
-    edge_sim = 1.0 - abs(fp_a["edge_density"] - fp_b["edge_density"])
-
-    # Color stats similarity (normalized Euclidean distance)
-    stats_a = np.array(fp_a["color_stats"], dtype=np.float64)
-    stats_b = np.array(fp_b["color_stats"], dtype=np.float64)
-    max_dist = np.sqrt(len(stats_a)) * 255.0  # theoretical max
-    dist = float(np.linalg.norm(stats_a - stats_b))
-    color_sim = 1.0 - (dist / max_dist) if max_dist > 0 else 1.0
-
-    # Weighted average. cv2.compareHist can return NaN on degenerate
-    # (e.g. all-zero) histograms; clamp would not catch it because
-    # NaN comparisons return False.
-    score = 0.6 * hist_sim + 0.2 * edge_sim + 0.2 * color_sim
-    if not math.isfinite(score):
-        return 0.0
-    return max(0.0, min(1.0, score))
-
-
-def scan_video_frames(
-    video_path: str,
-    region: dict[str, int] | None,
-    interval_seconds: float,
-    callback: ScanCallback,
-    *,
-    start_seconds: float = 0.0,
-    end_seconds: float | None = None,
-    fps: float = 0.0,
-    duration: float = 0.0,
-    fast_opts: dict[str, Any] | None = None,
-    cv_scale: float | None = None,
-) -> None:
-    """Iterate through video at interval, extract region, call callback.
-
-    The *callback* receives ``(timestamp_seconds, region_pixels)`` and may
-    return ``False`` to stop iteration early.  When *region* is ``None``
-    the full frame is passed (used by template detection).
-
-    When *fps* and *duration* are provided, skips internal metadata reads.
-    Uses sequential frame reading (grab/retrieve) for small intervals
-    to avoid expensive H.264 seeking.
-
-    *fast_opts* enables fast-scan optimizations when provided:
-    - ``phash_skip``: skip frames whose perceptual hash is unchanged
-    - ``max_region_dim``: downscale extracted region to this max dimension
-    """
-    full_frame = region is None
-    # Reject zero-dimension regions early so ffmpeg's crop filter doesn't
-    # produce empty frames downstream (which would crash cv2.cvtColor /
-    # cv2.GaussianBlur with an opaque shape mismatch).
-    if not full_frame and region is not None:
-        if region.get("w", 0) <= 0 or region.get("h", 0) <= 0:
-            utils.warning_print(
-                f"Skipping scan: region has zero width or height ({region})"
-            )
-            return
-    if cv_scale is None:
-        cv_scale = config.SCREENSPACE_CV_RESOLUTION_SCALE
-
-    # Extract frames via ffmpeg pipe (faster H.264 decoding, no cv2.VideoCapture)
-    if not _scan_via_ffmpeg_pipe(
-        video_path,
-        None if full_frame else region,
-        interval_seconds,
-        callback,
-        start_seconds=start_seconds,
-        end_seconds=end_seconds if end_seconds is not None else 0.0,
-        fps=fps,
-        duration=duration,
-        fast_opts=fast_opts,
-        full_frame=full_frame,
-        cv_scale=cv_scale,
-    ):
-        utils.warning_print(
-            f"ffmpeg pipe extraction failed for {Path(video_path).name}"
-        )
-
-
-def scan_video_full_frames(
-    video_path: str,
-    interval_seconds: float,
-    callback: ScanCallback,
-    *,
-    start_seconds: float = 0.0,
-    end_seconds: float | None = None,
-    fps: float = 0.0,
-    duration: float = 0.0,
-    fast_opts: dict[str, Any] | None = None,
-    cv_scale: float | None = None,
-) -> None:
-    """Like :func:`scan_video_frames` but passes the full frame (no region crop).
-
-    Thin wrapper that calls ``scan_video_frames`` with ``region=None``.
-    """
-    scan_video_frames(
-        video_path,
-        None,
-        interval_seconds,
-        callback,
-        start_seconds=start_seconds,
-        end_seconds=end_seconds,
-        fps=fps,
-        duration=duration,
-        fast_opts=fast_opts,
-        cv_scale=cv_scale,
-    )
-
-
-# ---------------------------------------------------------------------------
-# Batch frame extraction via ffmpeg pipe (experiment 2E)
-# ---------------------------------------------------------------------------
-
-
-def _ffmpeg_pipe_frames(
-    video_path: str,
-    interval_seconds: float,
-    *,
-    start_seconds: float = 0.0,
-    end_seconds: float = 0.0,
-    region: dict[str, int] | None = None,
-    frame_width: int = 0,
-    frame_height: int = 0,
-    max_dim: int = 0,
-    cv_scale: float = 1.0,
-) -> Iterator[tuple[float, np.ndarray]]:
-    """Yield ``(timestamp, frame)`` tuples extracted via an ffmpeg pipe.
-
-    Uses a single ffmpeg process with ``-f rawvideo`` piped to stdout,
-    which is typically faster than per-frame ``cv2.VideoCapture`` seeking
-    for H.264 content.
-
-    *region* applies an ffmpeg ``crop`` filter so only the ROI pixels are
-    decoded and transferred.  *max_dim* adds a ``scale`` filter to cap the
-    largest output dimension (useful for fast-scan downscaling).
-
-    The caller can stop iteration at any time (e.g. on cancel); the
-    ``finally`` block ensures the subprocess is cleaned up.
-    """
-    if not shutil.which("ffmpeg"):
-        return
-
-    # Determine output dimensions.
-    #
-    # `select` instead of `fps`: at interval N, fps=1/N picks the *last*
-    # source frame whose PTS rounds to each output slot — for 30 fps source
-    # that's ~0.5 s past the slot — and assigns it the slot's PTS in the
-    # output. The preview path seeks to that slot PTS and lands on a
-    # different source frame, hence the long-running click-vs-preview drift.
-    # `select` passes the chosen frame through with its original PTS, which
-    # showinfo (below) reports verbatim. Paired with `-fps_mode vfr` below to
-    # stop ffmpeg from duplicating frames to fill the source rate.
-    filters = [
-        f"select='isnan(prev_selected_t)+gte(t-prev_selected_t,{interval_seconds})'"
-    ]
-
-    if region:
-        filters.append(f"crop={region['w']}:{region['h']}:{region['x']}:{region['y']}")
-        out_w, out_h = region["w"], region["h"]
-    else:
-        out_w, out_h = frame_width, frame_height
-
-    if out_w <= 0 or out_h <= 0:
-        return
-
-    # Global CV resolution scale: applied after the region crop, before any
-    # max_dim cap. Skipped at 1.0 so default-config runs are byte-identical
-    # to pre-feature behavior.
-    if cv_scale > 0 and abs(cv_scale - 1.0) > 1e-6:
-        scaled_w = max(2, int(round(out_w * cv_scale)))
-        scaled_h = max(2, int(round(out_h * cv_scale)))
-        scaled_w += scaled_w % 2
-        scaled_h += scaled_h % 2
-        filters.append(f"scale={scaled_w}:{scaled_h}:flags=lanczos")
-        out_w, out_h = scaled_w, scaled_h
-
-    if max_dim > 0 and (out_w > max_dim or out_h > max_dim):
-        scale = max_dim / max(out_w, out_h)
-        out_w = int(out_w * scale)
-        out_h = int(out_h * scale)
-        # Ensure even dimensions for rawvideo
-        out_w += out_w % 2
-        out_h += out_h % 2
-        filters.append(f"scale={out_w}:{out_h}")
-
-    # showinfo last so its `pts_time:` lines correspond 1-to-1 with the
-    # rawvideo frames pushed to stdout. The drain thread below parses these
-    # and queues them so each yielded frame is tagged with its actual source
-    # PTS instead of a synthetic frame_idx*interval.
-    filters.append("showinfo")
-
-    cmd: list[str] = ["ffmpeg"]
-    if start_seconds > 0:
-        cmd += ["-ss", str(start_seconds)]
-    cmd += ["-i", video_path]
-    if end_seconds > start_seconds:
-        cmd += ["-t", str(end_seconds - start_seconds)]
-    cmd += [
-        "-vf",
-        ",".join(filters),
-        # `select` keeps source PTS, so without `vfr` ffmpeg pads the output
-        # back up to the source frame rate by duplicating each kept frame
-        # (~30 dupes per kept frame at 30 fps). vfr emits only the actual
-        # kept frames, one-to-one with the showinfo log lines.
-        "-fps_mode",
-        "vfr",
-        "-pix_fmt",
-        "bgr24",
-        "-f",
-        "rawvideo",
-        # showinfo emits at the `info` level; with `error` its lines never
-        # reach stderr and we'd have no PTS data to read.
-        "-loglevel",
-        "info",
-        "pipe:1",
-    ]
-
-    frame_size = out_w * out_h * 3
-    pts_re = re.compile(rb"pts_time:(\S+)")
-    pts_q: queue.Queue[float] = queue.Queue(maxsize=256)
-
-    proc = subprocess.Popen(cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-    assert proc.stdout is not None  # guaranteed by stdout=PIPE
-    assert proc.stderr is not None  # guaranteed by stderr=PIPE
-
-    # Daemon thread drains stderr so the OS buffer never blocks ffmpeg, and
-    # forwards every `pts_time:` line to the read loop as a float (seconds
-    # since the seek point).
-    def _drain_stderr() -> None:
-        assert proc.stderr is not None
-        for line in proc.stderr:
-            m = pts_re.search(line)
-            if m:
-                try:
-                    pts_q.put(float(m.group(1)))
-                except ValueError:
-                    pass
-
-    threading.Thread(target=_drain_stderr, daemon=True).start()
-
-    try:
-        while True:
-            raw = proc.stdout.read(frame_size)
-            if len(raw) < frame_size:
-                break
-            # Read-only view on the ffmpeg pipe bytes; callers treat frames as
-            # read-only and must .copy() if they retain past the next yield.
-            frame = np.frombuffer(raw, dtype=np.uint8).reshape((out_h, out_w, 3))
-            try:
-                relative_pts = pts_q.get(timeout=5.0)
-            except queue.Empty:
-                # Stderr stalled or showinfo missing — bail out rather than
-                # yield a frame with a misaligned synthetic timestamp.
-                break
-            actual_ts = start_seconds + relative_pts
-            if end_seconds > 0 and actual_ts > end_seconds:
-                break
-            yield (actual_ts, frame)
-    finally:
-        if proc.stdout:
-            proc.stdout.close()
-        utils.terminate_subprocess(proc)
-
-
-def _scan_via_ffmpeg_pipe(
-    video_path: str,
-    region: dict[str, int] | None,
-    interval_seconds: float,
-    callback: ScanCallback,
-    *,
-    start_seconds: float = 0.0,
-    end_seconds: float = 0.0,
-    fps: float = 0.0,
-    duration: float = 0.0,
-    fast_opts: dict[str, Any] | None = None,
-    full_frame: bool = False,
-    cv_scale: float = 1.0,
-) -> bool:
-    """Try to scan frames via ffmpeg pipe, calling *callback* for each.
-
-    Returns ``True`` if the ffmpeg path succeeded (caller should skip the
-    cv2 fallback), ``False`` if it failed and the caller should fall back
-    to cv2-based extraction.
-    """
-    if not shutil.which("ffmpeg"):
-        return False
-
-    props = video.probe_video_properties(video_path)
-    if not props:
-        return False
-    frame_width = props.get("width", 0)
-    frame_height = props.get("height", 0)
-    if frame_width <= 0 or frame_height <= 0:
-        return False
-
-    if end_seconds <= 0:
-        end_seconds = duration
-
-    # Fast-scan state
-    _phash_skip = bool(fast_opts and fast_opts.get("phash_skip"))
-    _max_dim = (fast_opts or {}).get("max_region_dim", 0)
-    _phash_thresh = (fast_opts or {}).get(
-        "phash_threshold", config.SCREENSPACE_FAST_SCAN_PHASH_THRESHOLD
-    )
-    _prev_phash: list["imagehash.ImageHash | None"] = [None]
-
-    pipe_region = None if full_frame else region
-    # For the pipe, push max_dim downscaling into ffmpeg when phash_skip is off.
-    # When phash_skip is on, we need the un-downscaled frame for hashing, so
-    # we downscale in Python after the hash check.
-    pipe_max_dim = _max_dim if (not _phash_skip and _max_dim > 0) else 0
-
-    try:
-        for ts, frame in _ffmpeg_pipe_frames(
-            video_path,
-            interval_seconds,
-            start_seconds=start_seconds,
-            end_seconds=end_seconds,
-            region=pipe_region,
-            frame_width=frame_width,
-            frame_height=frame_height,
-            max_dim=pipe_max_dim,
-            cv_scale=cv_scale,
-        ):
-            if _phash_skip:
-                fh = compute_phash(frame)
-                if _prev_phash[0] is not None and fh - _prev_phash[0] <= _phash_thresh:
-                    continue
-                _prev_phash[0] = fh
-                # Downscale after phash if needed
-                if _max_dim > 0:
-                    rh, rw = frame.shape[:2]
-                    if rh > _max_dim or rw > _max_dim:
-                        sc = _max_dim / max(rh, rw)
-                        frame = cv2.resize(
-                            frame,
-                            (int(rw * sc), int(rh * sc)),
-                            interpolation=cv2.INTER_AREA,
-                        )
-
-            result = callback(ts, frame)
-            if result is False:
-                break
-
-        return True  # ffmpeg pipe succeeded (even if video had 0 frames)
-    except Exception as exc:
-        utils.warning_print(f"ffmpeg pipe scan failed, falling back to cv2: {exc}")
-        return False
-
-
-def build_timelapse_command(
-    video_path: str,
-    region: dict[str, int],
-    speedup_factor: float,
-    output_path: str,
-    output_format: str = "mp4",
-    *,
-    start_seconds: float = 0.0,
-    end_seconds: float | None = None,
-    sample_interval: float = 0.0,
-) -> list[str]:
-    """Construct ffmpeg argv for a cropped timelapse.
-
-    *sample_interval* (seconds) controls frame sampling: when > 0, only one
-    frame per interval is kept before cropping and speed-up.  0 means every
-    frame is used (default).
-    """
-    x, y, w, h = region["x"], region["y"], region["w"], region["h"]
-    filters: list[str] = []
-    if sample_interval > 0:
-        filters.append(f"fps=1/{sample_interval}")
-    filters.append(f"crop={w}:{h}:{x}:{y}")
-    filters.append(f"setpts=PTS/{speedup_factor}")
-    vf = ",".join(filters)
-
-    cmd = [
-        "ffmpeg",
-        "-y",
-        "-loglevel",
-        config.FFMPEG_LOGLEVEL,
-    ]
-
-    if start_seconds > 0:
-        cmd += ["-ss", str(start_seconds)]
-
-    cmd += ["-i", video_path]
-
-    if end_seconds is not None and end_seconds > start_seconds:
-        cmd += ["-t", str(end_seconds - start_seconds)]
-
-    cmd += ["-vf", vf, "-an"]
-
-    if output_format == "gif":
-        cmd.extend(["-loop", "0"])
-    else:
-        cmd.extend(["-c:v", "libx264", "-preset", "fast", "-crf", "23"])
-
-    cmd.append(output_path)
-    return cmd
-
-
-def _probe_video_meta(video_path: str) -> tuple[float, float]:
-    """Return ``(fps, duration)`` via ffprobe."""
-    props = video.probe_video_properties(video_path)
-    if props and props.get("fps", 0) > 0 and props.get("duration", 0) > 0:
-        return (props["fps"], props["duration"])
-    return (0.0, 0.0)
-
-
-# ---------------------------------------------------------------------------
-# Analysis workflows
-# ---------------------------------------------------------------------------
-
-
-def scan_color(
-    video_path: str,
-    region: dict[str, int],
-    target_color: dict[str, float],
-    tolerance: dict[str, float],
-    interval_seconds: float = 0.0,
-    *,
-    start_seconds: float = 0.0,
-    end_seconds: float | None = None,
-    color_mode: str = "average",
-    min_coverage: float = 0.0,
-    on_progress: Callable[[float], None] | None = None,
-    cancel_flag: Callable[[], bool] | None = None,
-    on_result: Callable[[dict[str, Any]], None] | None = None,
-    fast_opts: dict[str, Any] | None = None,
-) -> list[dict[str, Any]]:
-    """Scan video for frames where the region matches a target color.
-
-    With ``color_mode="average"`` (default) a frame matches when the region's
-    *average* color is within tolerance. With ``color_mode="presence"`` a frame
-    matches when the target color appears anywhere in the region (per-pixel),
-    covering at least ``min_coverage`` of it.
-
-    Returns list of ``{start, end, duration}`` spans (consecutive matches
-    merged).
-    """
-    if interval_seconds <= 0:
-        interval_seconds = config.SCREENSPACE_DEFAULT_INTERVAL
-
-    vid_fps, vid_duration = _probe_video_meta(video_path)
-    if vid_fps <= 0:
-        return []
-
-    if end_seconds is None or end_seconds > vid_duration:
-        end_seconds = vid_duration
-    total_range = end_seconds - start_seconds
-
-    matches: list[float] = []
-
-    def _cb(ts: float, pixels: np.ndarray) -> bool | None:
-        if cancel_flag and cancel_flag():
-            return False
-        if color_mode == "presence":
-            matched, conf = color_present(pixels, target_color, tolerance, min_coverage)
-        else:
-            matched, conf = color_matches(pixels, target_color, tolerance)
-        if matched:
-            matches.append(ts)
-            if on_result:
-                on_result({"timestamp": ts, "_confidence": conf})
-        if on_progress and total_range > 0:
-            on_progress((ts - start_seconds) / total_range)
-        return None
-
-    scan_video_frames(
-        video_path,
-        region,
-        interval_seconds,
-        _cb,
-        start_seconds=start_seconds,
-        end_seconds=end_seconds,
-        fps=vid_fps,
-        duration=vid_duration,
-        fast_opts=fast_opts,
-    )
-
-    if on_progress:
-        on_progress(1.0)
-    return _merge_timestamp_spans(matches, interval_seconds)
-
-
-def scan_changes(
-    video_path: str,
-    region: dict[str, int],
-    threshold: float = 0.0,
-    interval_seconds: float = 0.0,
-    *,
-    noise_threshold: int = 0,
-    require_consecutive: int = 1,
-    start_seconds: float = 0.0,
-    end_seconds: float | None = None,
-    on_progress: Callable[[float], None] | None = None,
-    cancel_flag: Callable[[], bool] | None = None,
-    on_result: Callable[[dict[str, Any]], None] | None = None,
-    fast_opts: dict[str, Any] | None = None,
-) -> list[dict[str, Any]]:
-    """Scan video for change points in a region.
-
-    Returns list of ``{timestamp, magnitude}`` dicts.
-    """
-    if threshold <= 0:
-        threshold = config.SCREENSPACE_CHANGE_RATIO_THRESHOLD
-    if interval_seconds <= 0:
-        interval_seconds = config.SCREENSPACE_DEFAULT_INTERVAL
-    if noise_threshold <= 0:
-        noise_threshold = config.SCREENSPACE_NOISE_THRESHOLD
-
-    vid_fps, vid_duration = _probe_video_meta(video_path)
-    if vid_fps <= 0:
-        return []
-
-    if end_seconds is None or end_seconds > vid_duration:
-        end_seconds = vid_duration
-    total_range = end_seconds - start_seconds
-
-    results: list[dict[str, Any]] = []
-    prev_gray: list[np.ndarray | None] = [None]
-    k = config.SCREENSPACE_BLUR_KERNEL
-    morph_kernel = _morph_kernel(config.SCREENSPACE_MORPH_KERNEL)
-    buf = _ConsecutiveBuffer(require_consecutive)
-
-    def _cb(ts: float, pixels: np.ndarray) -> bool | None:
-        if cancel_flag and cancel_flag():
-            return False
-        curr_gray = cv2.cvtColor(
-            cv2.GaussianBlur(pixels, (k, k), 0), cv2.COLOR_BGR2GRAY
-        )
-        if prev_gray[0] is not None:
-            diff = cv2.absdiff(prev_gray[0], curr_gray)
-            _, mask = cv2.threshold(diff, noise_threshold, 255, cv2.THRESH_BINARY)
-            mask = cv2.morphologyEx(mask, cv2.MORPH_OPEN, morph_kernel)
-            mag = float(np.count_nonzero(mask)) / float(mask.size) if mask.size else 0.0
-            if mag >= threshold:
-                rd = {"timestamp": ts, "magnitude": round(mag, 4)}
-                emitted = buf.push(ts, rd)
-                if emitted is not None:
-                    results.append(emitted)
-                    if on_result:
-                        on_result(emitted)
-            else:
-                buf.reset()
-        prev_gray[0] = curr_gray
-        if on_progress and total_range > 0:
-            on_progress((ts - start_seconds) / total_range)
-        return None
-
-    scan_video_frames(
-        video_path,
-        region,
-        interval_seconds,
-        _cb,
-        start_seconds=start_seconds,
-        end_seconds=end_seconds,
-        fps=vid_fps,
-        duration=vid_duration,
-        fast_opts=fast_opts,
-    )
-
-    if on_progress:
-        on_progress(1.0)
-    return results
-
-
-def scan_similarity(
-    video_path: str,
-    region: dict[str, int],
-    reference_frame: np.ndarray,
-    threshold: float = 0.0,
-    interval_seconds: float = 0.0,
-    *,
-    start_seconds: float = 0.0,
-    end_seconds: float | None = None,
-    on_progress: Callable[[float], None] | None = None,
-    cancel_flag: Callable[[], bool] | None = None,
-    on_result: Callable[[dict[str, Any]], None] | None = None,
-    fast_opts: dict[str, Any] | None = None,
-) -> list[dict[str, Any]]:
-    """Find frames where region is similar to a reference.
-
-    Returns list of ``{timestamp, score}`` dicts, sorted by score
-    descending.
-    """
-    from skimage.metrics import structural_similarity as ssim
-
-    if threshold <= 0:
-        threshold = config.SCREENSPACE_SSIM_THRESHOLD
-    if interval_seconds <= 0:
-        interval_seconds = config.SCREENSPACE_DEFAULT_INTERVAL
-
-    vid_fps, vid_duration = _probe_video_meta(video_path)
-    if vid_fps <= 0:
-        return []
-
-    if end_seconds is None or end_seconds > vid_duration:
-        end_seconds = vid_duration
-    total_range = end_seconds - start_seconds
-
-    results: list[dict[str, Any]] = []
-    ref_phash = compute_phash(reference_frame)
-    phash_threshold = config.SCREENSPACE_PHASH_THRESHOLD
-
-    # Pre-resize and preprocess reference frame once
-    max_dim = 256
-    rh, rw = reference_frame.shape[:2]
-    if rh > max_dim or rw > max_dim:
-        scale = max_dim / max(rh, rw)
-        cmp_w, cmp_h = int(rw * scale), int(rh * scale)
-        ref_resized = cv2.resize(
-            reference_frame, (cmp_w, cmp_h), interpolation=cv2.INTER_AREA
-        )
-    else:
-        cmp_w, cmp_h = rw, rh
-        ref_resized = reference_frame
-    bk = config.SCREENSPACE_BLUR_KERNEL
-    ref_gray = cv2.cvtColor(
-        cv2.GaussianBlur(ref_resized, (bk, bk), 0), cv2.COLOR_BGR2GRAY
-    )
-
-    prev_skip_gray: list[np.ndarray | None] = [None]
-
-    def _cb(ts: float, pixels: np.ndarray) -> bool | None:
-        if cancel_flag and cancel_flag():
-            return False
-        # Static-frame skip
-        gray = cv2.cvtColor(pixels, cv2.COLOR_BGR2GRAY)
-        if prev_skip_gray[0] is not None:
-            if (
-                float(np.mean(cv2.absdiff(prev_skip_gray[0], gray)))
-                < config.SCREENSPACE_STATIC_FRAME_SKIP_THRESHOLD
-            ):
-                if on_progress and total_range > 0:
-                    on_progress((ts - start_seconds) / total_range)
-                return None
-        prev_skip_gray[0] = gray
-
-        frame_phash = compute_phash(pixels)
-        if ref_phash - frame_phash <= phash_threshold:
-            # Always resize candidate to match reference dimensions for SSIM
-            ph, pw = pixels.shape[:2]
-            if pw != cmp_w or ph != cmp_h:
-                cand = cv2.resize(pixels, (cmp_w, cmp_h), interpolation=cv2.INTER_AREA)
-            else:
-                cand = pixels
-            cand_gray = cv2.cvtColor(
-                cv2.GaussianBlur(cand, (bk, bk), 0), cv2.COLOR_BGR2GRAY
-            )
-            score = float(ssim(ref_gray, cand_gray))
-            if score >= threshold:
-                rd = {"timestamp": ts, "score": round(score, 4)}
-                results.append(rd)
-                if on_result:
-                    on_result(rd)
-        if on_progress and total_range > 0:
-            on_progress((ts - start_seconds) / total_range)
-        return None
-
-    scan_video_frames(
-        video_path,
-        region,
-        interval_seconds,
-        _cb,
-        start_seconds=start_seconds,
-        end_seconds=end_seconds,
-        fps=vid_fps,
-        duration=vid_duration,
-        fast_opts=fast_opts,
-    )
-
-    if on_progress:
-        on_progress(1.0)
-    results.sort(key=lambda r: r["score"], reverse=True)
-    return results
-
-
-def scan_text(
-    video_path: str,
-    region: dict[str, int],
-    search_string: str,
-    interval_seconds: float = 2.0,
-    *,
-    fuzzy_threshold: float = 0.0,
-    ocr_confidence_threshold: float | None = None,
-    ocr_preprocess: bool = False,
-    ocr_normalize: str = "off",
-    require_consecutive: int = 1,
-    languages: list[str] | None = None,
-    start_seconds: float = 0.0,
-    end_seconds: float | None = None,
-    on_progress: Callable[[float], None] | None = None,
-    cancel_flag: Callable[[], bool] | None = None,
-    on_result: Callable[[dict[str, Any]], None] | None = None,
-    fast_opts: dict[str, Any] | None = None,
-) -> list[dict[str, Any]]:
-    """Scan for text appearances in a region using EasyOCR.
-
-    EasyOCR is lazy-imported. Raises ``ImportError`` with install
-    instructions if missing.
-    """
-    if fuzzy_threshold <= 0:
-        fuzzy_threshold = config.SCREENSPACE_OCR_FUZZY_THRESHOLD
-    ocr_confidence_threshold = _effective_ocr_confidence_threshold(
-        ocr_confidence_threshold
-    )
-    utils.require_optional("easyocr", "text scan")
-    if languages is None:
-        languages = ["en"]
-
-    reader = _get_ocr_reader(languages)
-
-    vid_fps, vid_duration = _probe_video_meta(video_path)
-    if vid_fps <= 0:
-        return []
-
-    if end_seconds is None or end_seconds > vid_duration:
-        end_seconds = vid_duration
-    total_range = end_seconds - start_seconds
-
-    results: list[dict[str, Any]] = []
-    search_cmp = _normalize_ocr_text(search_string, ocr_normalize)
-    prev_gray: list[np.ndarray | None] = [None]
-    buf = _ConsecutiveBuffer(require_consecutive)
-
-    def _cb(ts: float, pixels: np.ndarray) -> bool | None:
-        if cancel_flag and cancel_flag():
-            return False
-        if _is_static_skip(
-            ts,
-            pixels,
-            prev_gray,
-            buf,
-            results,
-            on_result,
-            on_progress,
-            start_seconds,
-            total_range,
-        ):
-            return None
-        ocr_input = _preprocess_for_ocr(pixels) if ocr_preprocess else pixels
-        ocr_results = reader.readtext(ocr_input, detail=1)
-        matched_rd: dict[str, Any] | None = None
-        for _, text, conf in ocr_results:
-            if conf < ocr_confidence_threshold:
-                continue
-            ocr_cmp = _normalize_ocr_text(text, ocr_normalize)
-            ratio = difflib.SequenceMatcher(None, search_cmp, ocr_cmp).ratio()
-            if ratio >= fuzzy_threshold:
-                matched_rd = {
-                    "timestamp": ts,
-                    "text_found": text,
-                    "confidence": round(conf, 4),
-                }
-                break
-        if matched_rd is not None:
-            emitted = buf.push(ts, matched_rd)
-            if emitted is not None:
-                results.append(emitted)
-                if on_result:
-                    on_result(emitted)
-        else:
-            buf.reset()
-        if on_progress and total_range > 0:
-            on_progress((ts - start_seconds) / total_range)
-        return None
-
-    scan_video_frames(
-        video_path,
-        region,
-        interval_seconds,
-        _cb,
-        start_seconds=start_seconds,
-        end_seconds=end_seconds,
-        fps=vid_fps,
-        duration=vid_duration,
-        fast_opts=fast_opts,
-    )
-
-    if on_progress:
-        on_progress(1.0)
-    return results
-
-
-_NUMBERS_RE = re.compile(r"-?\d+(?:\.\d+)?")
-_VALID_OPERATORS = ("eq", "gt", "lt", "gte", "lte", "range")
-
-# EasyOCR character allowlist for the numbers tool. Numbers mode is digits-only,
-# so constraining recognition kills glyph confusions (O↔0, S↔5, l↔1) at the
-# source. Only passed for the default English reader — some language combos
-# reject ``allowlist``. Mirrors what the downstream parser accepts (``-``, ``.``,
-# ``,`` thousands separators, digits).
-_OCR_NUMBER_ALLOWLIST = "0123456789.,-"
-# integers_only narrows the allowlist further: dropping ``.,-`` stops a separator
-# or sign glyph from surviving OCR as a digit and inflating the parsed value.
-# Use for whole-number HUD targets (scores, counts) where decimals never appear.
-_OCR_DIGITS_ONLY_ALLOWLIST = "0123456789"
-
-
-def _number_matches(
-    value: float,
-    operator: str,
-    target_value: float = 0,
-    range_min: float | None = None,
-    range_max: float | None = None,
-) -> bool:
-    """Check if *value* satisfies the given numeric comparison."""
-    if operator == "eq":
-        return value == target_value
-    elif operator == "gt":
-        return value > target_value
-    elif operator == "lt":
-        return value < target_value
-    elif operator == "gte":
-        return value >= target_value
-    elif operator == "lte":
-        return value <= target_value
-    elif operator == "range":
-        return (
-            range_min is not None
-            and range_max is not None
-            and range_min <= value <= range_max
-        )
-    return False
-
-
-def scan_numbers(
-    video_path: str,
-    region: dict[str, int],
-    operator: str,
-    target_value: float = 0,
-    interval_seconds: float = 2.0,
-    *,
-    range_min: float | None = None,
-    range_max: float | None = None,
-    ocr_confidence_threshold: float | None = None,
-    ocr_preprocess: bool = False,
-    integers_only: bool = False,
-    require_consecutive: int = 1,
-    languages: list[str] | None = None,
-    start_seconds: float = 0.0,
-    end_seconds: float | None = None,
-    on_progress: Callable[[float], None] | None = None,
-    cancel_flag: Callable[[], bool] | None = None,
-    on_result: Callable[[dict[str, Any]], None] | None = None,
-    fast_opts: dict[str, Any] | None = None,
-) -> list[dict[str, Any]]:
-    """Scan for numeric values in a region and apply a comparison.
-
-    Uses EasyOCR to detect text, parses numbers from it, and returns
-    timestamps where the detected number satisfies the comparison.
-    """
-    if operator not in _VALID_OPERATORS:
-        raise ValueError(
-            f"Unknown operator '{operator}'. Must be one of: {', '.join(_VALID_OPERATORS)}"
-        )
-
-    ocr_confidence_threshold = _effective_ocr_confidence_threshold(
-        ocr_confidence_threshold
-    )
-    utils.require_optional("easyocr", "numbers scan")
-    if languages is None:
-        languages = ["en"]
-
-    reader = _get_ocr_reader(languages)
-
-    vid_fps, vid_duration = _probe_video_meta(video_path)
-    if vid_fps <= 0:
-        return []
-
-    if end_seconds is None or end_seconds > vid_duration:
-        end_seconds = vid_duration
-    total_range = end_seconds - start_seconds
-
-    results: list[dict[str, Any]] = []
-    prev_gray: list[np.ndarray | None] = [None]
-    # Hoisted out of the per-frame callback: constrain English OCR to digits.
-    ocr_kwargs: dict[str, Any] = {"detail": 1}
-    if languages == ["en"]:
-        ocr_kwargs["allowlist"] = (
-            _OCR_DIGITS_ONLY_ALLOWLIST if integers_only else _OCR_NUMBER_ALLOWLIST
-        )
-    buf = _ConsecutiveBuffer(require_consecutive)
-
-    def _cb(ts: float, pixels: np.ndarray) -> bool | None:
-        if cancel_flag and cancel_flag():
-            return False
-        if _is_static_skip(
-            ts,
-            pixels,
-            prev_gray,
-            buf,
-            results,
-            on_result,
-            on_progress,
-            start_seconds,
-            total_range,
-        ):
-            return None
-        ocr_input = _preprocess_for_ocr(pixels) if ocr_preprocess else pixels
-        ocr_results = reader.readtext(ocr_input, **ocr_kwargs)
-        matched_rd: dict[str, Any] | None = None
-        for _, text, conf in ocr_results:
-            if conf < ocr_confidence_threshold:
-                continue
-            cleaned = text.replace(",", "")
-            for match in _NUMBERS_RE.findall(cleaned):
-                num = float(match)
-                if _number_matches(num, operator, target_value, range_min, range_max):
-                    matched_rd = {
-                        "timestamp": ts,
-                        "number_found": num,
-                        "confidence": round(conf, 4),
-                    }
-                    break
-            if matched_rd is not None:
-                break
-        if matched_rd is not None:
-            emitted = buf.push(ts, matched_rd)
-            if emitted is not None:
-                results.append(emitted)
-                if on_result:
-                    on_result(emitted)
-        else:
-            buf.reset()
-        if on_progress and total_range > 0:
-            on_progress((ts - start_seconds) / total_range)
-        return None
-
-    scan_video_frames(
-        video_path,
-        region,
-        interval_seconds,
-        _cb,
-        start_seconds=start_seconds,
-        end_seconds=end_seconds,
-        fps=vid_fps,
-        duration=vid_duration,
-        fast_opts=fast_opts,
-    )
-
-    if on_progress:
-        on_progress(1.0)
-    return results
-
-
-def generate_timelapse(
-    video_path: str,
-    region: dict[str, int],
-    speedup_factor: float,
-    output_path: str,
-    output_format: str = "mp4",
-    *,
-    start_seconds: float = 0.0,
-    end_seconds: float | None = None,
-    sample_interval: float = 0.0,
-    on_progress: Callable[[float], None] | None = None,
-    cancel_flag: Callable[[], bool] | None = None,
-) -> str | None:
-    """Generate a cropped timelapse via ffmpeg.
-
-    *sample_interval* controls frame sampling (seconds between captured
-    frames).  0 means every frame is used.
-
-    Reports encoding progress through *on_progress* by parsing ffmpeg's
-    ``-progress`` output.  Checks *cancel_flag* periodically to allow
-    early termination.
-
-    Returns output file path on success, ``None`` on failure.
-    """
-    cmd = build_timelapse_command(
-        video_path,
-        region,
-        speedup_factor,
-        output_path,
-        output_format,
-        start_seconds=start_seconds,
-        end_seconds=end_seconds,
-        sample_interval=sample_interval,
-    )
-
-    # Estimate output duration for progress calculation.
-    input_duration = 0.0
-    if end_seconds is not None and end_seconds > start_seconds:
-        input_duration = end_seconds - start_seconds
-    else:
-        _, vid_dur = _probe_video_meta(video_path)
-        if vid_dur > 0:
-            input_duration = max(vid_dur - start_seconds, 0.0)
-
-    expected_out_us = (
-        (input_duration / speedup_factor * 1_000_000) if input_duration > 0 else 0
-    )
-
-    # Use Popen with -progress to get real-time encoding updates
-    cmd += ["-progress", "pipe:1"]
-    try:
-        # stderr → DEVNULL: only stdout (progress lines) is read, so a PIPE'd
-        # stderr could fill its 64 KB OS buffer and deadlock ffmpeg.
-        proc = subprocess.Popen(cmd, stdout=subprocess.PIPE, stderr=subprocess.DEVNULL)
-        assert proc.stdout is not None  # guaranteed by stdout=PIPE
-    except (FileNotFoundError, OSError):
-        return None
-
-    if on_progress:
-        on_progress(0.0)
-
-    try:
-        for line in proc.stdout:
-            text = line.decode("utf-8", errors="replace").strip()
-            if text.startswith("out_time_us=") and expected_out_us > 0 and on_progress:
-                try:
-                    us = int(text.split("=", 1)[1])
-                    on_progress(min(us / expected_out_us, 0.99))
-                except (ValueError, ZeroDivisionError):
-                    pass
-            if cancel_flag and cancel_flag():
-                utils.terminate_subprocess(proc)
-                return None
-    finally:
-        if proc.stdout:
-            proc.stdout.close()
-        proc.wait()
-
-    if on_progress:
-        on_progress(1.0)
-    if proc.returncode != 0:
-        return None
-    return output_path
-
-
-def scan_template(
-    video_path: str,
-    region: dict[str, int],
-    template_image: np.ndarray,
-    threshold: float = 0.0,
-    interval_seconds: float = 0.0,
-    *,
-    template_mask: np.ndarray | None = None,
-    template_scale: float = 1.0,
-    start_seconds: float = 0.0,
-    end_seconds: float | None = None,
-    on_progress: Callable[[float], None] | None = None,
-    cancel_flag: Callable[[], bool] | None = None,
-    on_result: Callable[[dict[str, Any]], None] | None = None,
-    fast_opts: dict[str, Any] | None = None,
-) -> list[dict[str, Any]]:
-    """Scan video for frames containing the template image.
-
-    *template_scale* resizes the uploaded template before matching
-    (e.g. ``0.5`` for a template captured at 2x the in-video scale).  An
-    optional *template_mask* restricts matching to non-transparent regions
-    of an uploaded PNG.
-
-    Returns list of ``{timestamp, matches, best_score, match_count}`` dicts.
-    """
-    if threshold <= 0:
-        threshold = config.SCREENSPACE_TEMPLATE_MATCH_THRESHOLD
-    if interval_seconds <= 0:
-        interval_seconds = config.SCREENSPACE_DEFAULT_INTERVAL
-
-    vid_fps, vid_duration = _probe_video_meta(video_path)
-    if vid_fps <= 0:
-        return []
-
-    if end_seconds is None or end_seconds > vid_duration:
-        end_seconds = vid_duration
-    total_range = end_seconds - start_seconds
-
-    results: list[dict[str, Any]] = []
-
-    _tmpl_downscale = bool(fast_opts and fast_opts.get("template_downscale"))
-
-    scaled_template, scaled_mask = _scale_template(
-        template_image, template_mask, template_scale
-    )
-
-    # Hoist the constant template prep (blur + grayscale + variance check)
-    # out of the per-frame callback. A degenerate template yields [] every
-    # frame, so bail early before even opening the ffmpeg pipe.
-    _prepared = _prepare_template(scaled_template, scaled_mask)
-    if _prepared[2]:  # degenerate template
-        if on_progress:
-            on_progress(1.0)
-        return results
-
-    _nms_overlap = config.SCREENSPACE_TEMPLATE_NMS_OVERLAP
-    # Frame is already at cv_scale * original due to ffmpeg scaling. We need
-    # to undo both the user-set cv_scale and the fast-scan internal 2x
-    # downscale to report match boxes in original-frame pixels.
-    _cv_scale = (
-        config.SCREENSPACE_CV_RESOLUTION_SCALE
-        if config.SCREENSPACE_CV_RESOLUTION_SCALE > 0
-        else 1.0
-    )
-
-    def _cb(ts: float, frame: np.ndarray) -> bool | None:
-        if cancel_flag and cancel_flag():
-            return False
-        work_frame = frame
-        scale_back = 1
-        if _tmpl_downscale:
-            fh, fw = work_frame.shape[:2]
-            nw, nh = fw // 2, fh // 2
-            if nw > 0 and nh > 0:
-                work_frame = cv2.resize(
-                    work_frame, (nw, nh), interpolation=cv2.INTER_AREA
-                )
-                scale_back = 2
-        matches = _match_template_prepared(
-            work_frame, _prepared, threshold, _nms_overlap
-        )
-        if matches:
-            # Undo fast-scan downscale, then undo cv_scale, so reported
-            # coords are in the original (un-scaled) frame coordinate space.
-            inv = scale_back / _cv_scale
-            if abs(inv - 1.0) > 1e-6:
-                for m in matches:
-                    m["x"] = int(round(m["x"] * inv))
-                    m["y"] = int(round(m["y"] * inv))
-                    m["w"] = int(round(m["w"] * inv))
-                    m["h"] = int(round(m["h"] * inv))
-            best = max(m["score"] for m in matches)
-            rd = {
-                "timestamp": ts,
-                "matches": matches,
-                "best_score": round(best, 4),
-                "match_count": len(matches),
-            }
-            results.append(rd)
-            if on_result:
-                on_result(
-                    {
-                        "timestamp": ts,
-                        "best_score": rd["best_score"],
-                        "match_count": rd["match_count"],
-                    }
-                )
-        if on_progress and total_range > 0:
-            on_progress((ts - start_seconds) / total_range)
-        return None
-
-    scan_video_full_frames(
-        video_path,
-        interval_seconds,
-        _cb,
-        start_seconds=start_seconds,
-        end_seconds=end_seconds,
-        fps=vid_fps,
-        duration=vid_duration,
-        fast_opts=fast_opts,
-    )
-
-    if on_progress:
-        on_progress(1.0)
-    return results
-
-
-def scan_flow(
-    video_path: str,
-    region: dict[str, int],
-    magnitude_threshold: float = 0.0,
-    interval_seconds: float = 0.0,
-    *,
-    require_consecutive: int = 1,
-    start_seconds: float = 0.0,
-    end_seconds: float | None = None,
-    on_progress: Callable[[float], None] | None = None,
-    cancel_flag: Callable[[], bool] | None = None,
-    on_result: Callable[[dict[str, Any]], None] | None = None,
-    fast_opts: dict[str, Any] | None = None,
-) -> list[dict[str, Any]]:
-    """Scan video for motion in a region using dense optical flow.
-
-    Returns list of ``{timestamp, magnitude, angle}`` dicts where
-    magnitude exceeds *magnitude_threshold*.
-    """
-    if magnitude_threshold <= 0:
-        magnitude_threshold = config.SCREENSPACE_FLOW_MAGNITUDE_THRESHOLD
-    if interval_seconds <= 0:
-        interval_seconds = config.SCREENSPACE_DEFAULT_INTERVAL
-
-    vid_fps, vid_duration = _probe_video_meta(video_path)
-    if vid_fps <= 0:
-        return []
-
-    if end_seconds is None or end_seconds > vid_duration:
-        end_seconds = vid_duration
-    total_range = end_seconds - start_seconds
-
-    results: list[dict[str, Any]] = []
-    prev_gray: list[np.ndarray | None] = [None]
-    buf = _ConsecutiveBuffer(require_consecutive)
-
-    def _cb(ts: float, pixels: np.ndarray) -> bool | None:
-        if cancel_flag and cancel_flag():
-            return False
-        curr_gray = cv2.cvtColor(pixels, cv2.COLOR_BGR2GRAY)
-        if prev_gray[0] is not None:
-            flow_result = compute_optical_flow(
-                prev_gray[0], curr_gray, return_grid=True
-            )
-            if flow_result["magnitude"] >= magnitude_threshold:
-                rd: dict[str, Any] = {
-                    "timestamp": ts,
-                    "magnitude": flow_result["magnitude"],
-                    "angle": flow_result["angle"],
-                    "flow_grid": flow_result.get("flow_grid", []),
-                }
-                emitted = buf.push(ts, rd)
-                if emitted is not None:
-                    results.append(emitted)
-                    if on_result:
-                        # Keep incremental updates lightweight (no grid)
-                        on_result(
-                            {
-                                "timestamp": emitted["timestamp"],
-                                "magnitude": emitted["magnitude"],
-                                "angle": emitted["angle"],
-                            }
-                        )
-            else:
-                buf.reset()
-        prev_gray[0] = curr_gray
-        if on_progress and total_range > 0:
-            on_progress((ts - start_seconds) / total_range)
-        return None
-
-    scan_video_frames(
-        video_path,
-        region,
-        interval_seconds,
-        _cb,
-        start_seconds=start_seconds,
-        end_seconds=end_seconds,
-        fps=vid_fps,
-        duration=vid_duration,
-        fast_opts=fast_opts,
-    )
-
-    if on_progress:
-        on_progress(1.0)
-    return results
-
-
-def scan_scene(
-    video_path: str,
-    region: dict[str, int],
-    reference_scenes: list[dict[str, Any]],
-    threshold: float = 0.0,
-    interval_seconds: float = 0.0,
-    *,
-    start_seconds: float = 0.0,
-    end_seconds: float | None = None,
-    on_progress: Callable[[float], None] | None = None,
-    cancel_flag: Callable[[], bool] | None = None,
-    on_result: Callable[[dict[str, Any]], None] | None = None,
-    fast_opts: dict[str, Any] | None = None,
-) -> list[dict[str, Any]]:
-    """Classify frames by similarity to reference scene fingerprints.
-
-    *reference_scenes* is a list of ``{name: str, frame: np.ndarray}``
-    dicts, optionally with a per-scene ``threshold`` float.  Each
-    frame's region is fingerprinted and compared against all references;
-    the best match above its threshold is reported.
-
-    Returns list of ``{timestamp, scene_name, score}`` dicts.
-    """
-    default_threshold = (
-        threshold if threshold > 0 else config.SCREENSPACE_SCENE_SIMILARITY_THRESHOLD
-    )
-    if interval_seconds <= 0:
-        interval_seconds = config.SCREENSPACE_DEFAULT_INTERVAL
-
-    # Pre-compute fingerprints for reference scenes (with per-scene thresholds)
-    ref_fps: list[tuple[str, dict[str, Any], float]] = []
-    for ref in reference_scenes:
-        fp = compute_scene_fingerprint(ref["frame"])
-        ref_thresh = float(ref.get("threshold", default_threshold))
-        ref_fps.append((ref["name"], fp, ref_thresh))
-
-    if not ref_fps:
-        return []
-
-    vid_fps, vid_duration = _probe_video_meta(video_path)
-    if vid_fps <= 0:
-        return []
-
-    if end_seconds is None or end_seconds > vid_duration:
-        end_seconds = vid_duration
-    total_range = end_seconds - start_seconds
-
-    results: list[dict[str, Any]] = []
-    prev_skip_gray: list[np.ndarray | None] = [None]
-
-    def _cb(ts: float, pixels: np.ndarray) -> bool | None:
-        if cancel_flag and cancel_flag():
-            return False
-
-        # Static-frame skip (same pattern as similarity scan)
-        curr_gray = cv2.cvtColor(pixels, cv2.COLOR_BGR2GRAY).astype(np.float32)
-        if prev_skip_gray[0] is not None:
-            if (
-                abs(float(np.mean(curr_gray)) - float(np.mean(prev_skip_gray[0])))
-                < config.SCREENSPACE_STATIC_FRAME_SKIP_THRESHOLD
-            ):
-                if on_progress and total_range > 0:
-                    on_progress((ts - start_seconds) / total_range)
-                return None
-        prev_skip_gray[0] = curr_gray
-
-        fp = compute_scene_fingerprint(pixels)
-        best_name = ""
-        best_score = 0.0
-        best_thresh = default_threshold
-        for ref_name, ref_fp, ref_thresh in ref_fps:
-            score = compare_scene_fingerprints(fp, ref_fp)
-            if score > best_score:
-                best_score = score
-                best_name = ref_name
-                best_thresh = ref_thresh
-
-        if best_score >= best_thresh:
-            rd = {
-                "timestamp": ts,
-                "scene_name": best_name,
-                "score": round(best_score, 4),
-            }
-            results.append(rd)
-            if on_result:
-                on_result(rd)
-        if on_progress and total_range > 0:
-            on_progress((ts - start_seconds) / total_range)
-        return None
-
-    scan_video_frames(
-        video_path,
-        region,
-        interval_seconds,
-        _cb,
-        start_seconds=start_seconds,
-        end_seconds=end_seconds,
-        fps=vid_fps,
-        duration=vid_duration,
-        fast_opts=fast_opts,
-    )
-
-    if on_progress:
-        on_progress(1.0)
-    return results
-
-
-def scan_inactivity(
-    video_path: str,
-    region: dict[str, int],
-    threshold: int = 0,
-    min_duration: float = 0.0,
-    interval_seconds: float = 0.0,
-    *,
-    start_seconds: float = 0.0,
-    end_seconds: float | None = None,
-    on_progress: Callable[[float], None] | None = None,
-    cancel_flag: Callable[[], bool] | None = None,
-    on_result: Callable[[dict[str, Any]], None] | None = None,
-    fast_opts: dict[str, Any] | None = None,
-) -> list[dict[str, Any]]:
-    """Scan video for inactivity spans (near-duplicate consecutive frames).
-
-    Uses perceptual hashing to compare consecutive frames.  When frames
-    remain similar (Hamming distance <= *threshold*) for at least
-    *min_duration* seconds, a span result is emitted.
-
-    Returns list of ``{start, end, duration, avg_distance}`` dicts.
-    """
-    if threshold <= 0:
-        threshold = config.SCREENSPACE_INACTIVITY_PHASH_THRESHOLD
-    if min_duration <= 0:
-        min_duration = config.SCREENSPACE_INACTIVITY_MIN_DURATION
-    if interval_seconds <= 0:
-        interval_seconds = config.SCREENSPACE_DEFAULT_INTERVAL
-
-    vid_fps, vid_duration = _probe_video_meta(video_path)
-    if vid_fps <= 0:
-        return []
-
-    if end_seconds is None or end_seconds > vid_duration:
-        end_seconds = vid_duration
-    total_range = end_seconds - start_seconds
-
-    results: list[dict[str, Any]] = []
-    prev_hash: list["imagehash.ImageHash | None"] = [None]
-    span_start: list[float | None] = [None]
-    span_distances: list[list[int]] = [[]]
-    last_ts: list[float] = [start_seconds]
-
-    def _flush_span(span_end: float) -> None:
-        if span_start[0] is not None:
-            dur = span_end - span_start[0]
-            if dur >= min_duration:
-                dists = span_distances[0]
-                avg_dist = sum(dists) / len(dists) if dists else 0.0
-                rd = {
-                    "start": round(span_start[0], 2),
-                    "end": round(span_end, 2),
-                    "duration": round(dur, 2),
-                    "avg_distance": round(avg_dist, 2),
-                }
-                results.append(rd)
-                if on_result:
-                    on_result(rd)
-        span_start[0] = None
-        span_distances[0] = []
-
-    def _cb(ts: float, pixels: np.ndarray) -> bool | None:
-        if cancel_flag and cancel_flag():
-            return False
-
-        curr_hash = compute_phash(pixels)
-        last_ts[0] = ts
-
-        if prev_hash[0] is not None:
-            dist = int(curr_hash - prev_hash[0])
-            if dist <= threshold:
-                # Frame is similar — extend or start span
-                if span_start[0] is None:
-                    span_start[0] = ts - interval_seconds
-                span_distances[0].append(dist)
-            else:
-                # Frame changed — flush any active span
-                _flush_span(ts - interval_seconds)
-        prev_hash[0] = curr_hash
-
-        if on_progress and total_range > 0:
-            on_progress((ts - start_seconds) / total_range)
-        return None
-
-    scan_video_frames(
-        video_path,
-        region,
-        interval_seconds,
-        _cb,
-        start_seconds=start_seconds,
-        end_seconds=end_seconds,
-        fps=vid_fps,
-        duration=vid_duration,
-        fast_opts=fast_opts,
-    )
-
-    # Flush final span if video ended during an inactive period
-    _flush_span(last_ts[0])
-
-    if on_progress:
-        on_progress(1.0)
-    return results
-
-
-def scan_boundaries(
-    video_path: str,
-    region: dict[str, int] | None = None,
-    threshold: int = 0,
-    min_gap: float = 0.0,
-    interval_seconds: float = 0.0,
-    *,
-    start_seconds: float = 0.0,
-    end_seconds: float | None = None,
-    on_progress: Callable[[float], None] | None = None,
-    cancel_flag: Callable[[], bool] | None = None,
-    on_result: Callable[[dict[str, Any]], None] | None = None,
-    fast_opts: dict[str, Any] | None = None,
-) -> list[dict[str, Any]]:
-    """Scan the full frame for scene boundaries (large frame-to-frame jumps).
-
-    The inverse of :func:`scan_inactivity`: where inactivity looks for phash
-    *plateaus*, boundaries look for *spikes*.  A boundary fires when the
-    perceptual-hash Hamming distance to the previous sampled frame is at least
-    *threshold*, debounced so that no two boundaries land within *min_gap*
-    seconds (the first spike of a sustained high-distance run is the boundary,
-    which keeps camera-heavy action sequences from becoming boundary storms).
-
-    Region is ignored (full-frame only); the parameter exists for signature
-    parity with the other scanners.  Returns a list of
-    ``{timestamp, distance, _confidence}`` dicts.
-    """
-    if threshold <= 0:
-        threshold = config.SCREENSPACE_BOUNDARY_PHASH_THRESHOLD
-    if min_gap <= 0:
-        min_gap = config.SCREENSPACE_BOUNDARY_MIN_GAP_SECONDS
-    if interval_seconds <= 0:
-        interval_seconds = config.SCREENSPACE_BOUNDARY_INTERVAL
-
-    vid_fps, vid_duration = _probe_video_meta(video_path)
-    if vid_fps <= 0:
-        return []
-
-    if end_seconds is None or end_seconds > vid_duration:
-        end_seconds = vid_duration
-    total_range = end_seconds - start_seconds
-
-    # Downscale frames at the ffmpeg pipe for cheap hashing. We pass only
-    # ``max_region_dim`` (no ``phash_skip``) so the pipe downsizes without
-    # dropping frames — this scanner runs its own phash on every sample.
-    boundary_opts = dict(fast_opts or {})
-    boundary_opts.setdefault("max_region_dim", config.SCREENSPACE_BOUNDARY_HASH_DIM)
-    boundary_opts.pop("phash_skip", None)
-
-    results: list[dict[str, Any]] = []
-    prev_hash: list["imagehash.ImageHash | None"] = [None]
-    last_boundary_ts: list[float | None] = [None]
-    eps = config.SCREENSPACE_BOUNDARY_CONFIDENCE_EPSILON
-
-    def _cb(ts: float, pixels: np.ndarray) -> bool | None:
-        if cancel_flag and cancel_flag():
-            return False
-
-        curr_hash = compute_phash(pixels)
-        if prev_hash[0] is not None:
-            dist = int(curr_hash - prev_hash[0])
-            within_gap = (
-                last_boundary_ts[0] is not None and ts - last_boundary_ts[0] < min_gap
-            )
-            if dist >= threshold and not within_gap:
-                conf = 1.0 if threshold <= 0 else (dist - threshold) / float(threshold)
-                conf = max(eps, min(conf, 1.0))
-                rd = {
-                    "timestamp": round(ts, 2),
-                    "distance": dist,
-                    "_confidence": round(conf, 4),
-                }
-                results.append(rd)
-                if on_result:
-                    on_result(rd)
-                last_boundary_ts[0] = ts
-        prev_hash[0] = curr_hash
-
-        if on_progress and total_range > 0:
-            on_progress((ts - start_seconds) / total_range)
-        return None
-
-    scan_video_full_frames(
-        video_path,
-        interval_seconds,
-        _cb,
-        start_seconds=start_seconds,
-        end_seconds=end_seconds,
-        fps=vid_fps,
-        duration=vid_duration,
-        fast_opts=boundary_opts,
-    )
-
-    if on_progress:
-        on_progress(1.0)
-    return results
-
-
-# ---------------------------------------------------------------------------
-# Multitool: per-frame evaluation and multi-factor scan
-# ---------------------------------------------------------------------------
-
-
-def _extract_confidence(tool_type: str, result: dict[str, Any]) -> float:
-    """Extract a normalized [0, 1] confidence from a tool-specific result dict."""
-    if tool_type == "color":
-        return result.get("_confidence", 1.0)
-    elif tool_type == "change":
-        return min(result.get("magnitude", 0.0), 1.0)
-    elif tool_type == "similarity":
-        return result.get("score", 0.0)
-    elif tool_type == "text":
-        return result.get("confidence", 0.0)
-    elif tool_type == "numbers":
-        return result.get("confidence", 1.0)
-    elif tool_type == "template":
-        return result.get("best_score", 0.0)
-    elif tool_type == "flow":
-        return min(result.get("magnitude", 0.0) / 10.0, 1.0)
-    elif tool_type == "scene":
-        return result.get("score", 0.0)
-    elif tool_type == "multitool":
-        return result.get("min_confidence", 0.0)
-    elif tool_type == "inactivity":
-        if "_confidence" in result:
-            return float(result["_confidence"])
-        return min(result.get("duration", 0.0) / 30.0, 1.0)
-    elif tool_type == "boundary":
-        if "_confidence" in result:
-            return float(result["_confidence"])
-        thr = config.SCREENSPACE_BOUNDARY_PHASH_THRESHOLD
-        dist = result.get("distance", 0.0)
-        if thr <= 0:
-            return 1.0
-        return max(0.0, min((dist - thr) / float(thr), 1.0))
-    return 1.0
-
-
-def check_frame_for_tool(
-    frame: np.ndarray,
-    prev_frame: np.ndarray | None,
-    region: dict[str, int],
-    tool_type: str,
-    parameters: dict[str, Any],
-) -> tuple[bool, dict[str, Any] | None]:
-    """Evaluate whether a single frame passes a tool's criteria.
-
-    Used by :func:`scan_multitool` for steps 1+ in the chain.  Returns
-    ``(passed, result_dict)`` where *result_dict* contains tool-specific
-    metadata when the check passes, or ``None`` when it does not.
-
-    For **change** and **flow** tools *prev_frame* is required (the frame
-    immediately before the candidate timestamp).  If it is ``None`` the
-    check is skipped (returns ``(False, None)``).
-
-    Degenerate regions (width or height ≤ 0, e.g. a 1-px user draw rounded
-    to zero pixels on a small preview) are treated as a non-match here so
-    downstream cv2 ops never see empty arrays.  Template is exempt: it matches
-    against the full frame and ignores ``region``, so an uploaded template step
-    with no region (zero-size region_coords) is valid — mirrors
-    :func:`score_frame_for_tool`.
-    """
-    if tool_type != "template" and (region.get("w", 0) <= 0 or region.get("h", 0) <= 0):
-        return False, None
-    tool = TOOLS.get(tool_type)
-    if tool is None:
-        return False, None
-    return tool.check_frame(frame, prev_frame, region, parameters)
-
-
-def _score_result(
-    score_key: str, passed: bool, detail: dict[str, Any] | None
-) -> dict[str, Any]:
-    """Wrap a ``check_frame``-style ``(passed, detail)`` into a calibration score.
-
-    Reads the threshold-independent scalar at ``detail[score_key]``. Returns a
-    ``not_evaluable`` status when the frame could not be scored (missing detail,
-    empty/missing score key, or a non-finite scalar from numpy/OpenCV).
-    """
-    if detail is None or not score_key:
-        return {"status": "not_evaluable"}
-    raw = detail.get(score_key)
-    if raw is None:
-        return {"status": "not_evaluable"}
-    score = float(raw)
-    if not math.isfinite(score):
-        return {"status": "not_evaluable"}
-    return {"status": "ok", "score": score, "passed": bool(passed), "detail": detail}
-
-
-def score_frame_for_tool(
-    tool_type: str,
-    frame: np.ndarray,
-    prev_frame: np.ndarray | None,
-    region: dict[str, int],
-    params: dict[str, Any],
-    *,
-    ocr_reader: "Callable[[str, dict[str, int], dict[str, Any]], list[Any]] | None" = None,
-) -> dict[str, Any]:
-    """Score a single frame against one tool's parameters for pin calibration.
-
-    Mirrors :func:`check_frame_for_tool` but returns the threshold-independent
-    scalar (``{status, score, passed, detail}``) instead of a boolean. Degenerate
-    regions and unknown / non-scorable tools (timelapse, multitool) return
-    ``not_evaluable``.
-
-    ``ocr_reader`` (text/numbers only) supplies cached EasyOCR readings keyed per
-    pin so fuzzy/confidence changes re-score without re-running OCR; when absent
-    the tool runs OCR live through its ``check_frame``.
-    """
-    # Template matches against the full frame and ignores ``region``, so a
-    # zero-size region — the case when an uploaded template scans the whole
-    # frame with no region_ref — is valid for it. Every other tool crops the
-    # region, where an empty crop would break downstream cv2 ops.
-    if tool_type != "template" and (region.get("w", 0) <= 0 or region.get("h", 0) <= 0):
-        return {"status": "not_evaluable"}
-    tool = TOOLS.get(tool_type)
-    if tool is None or not tool.score_key:
-        return {"status": "not_evaluable"}
-    if ocr_reader is not None and tool_type in ("text", "numbers"):
-        if tool_type == "text" and not params.get("search_string", ""):
-            return {"status": "not_evaluable"}
-        readings = ocr_reader(tool_type, region, params)
-        if tool_type == "text":
-            passed, detail = _score_text_readings(readings, params)
-        else:
-            passed, detail = _score_numbers_readings(readings, params)
-        return _score_result(tool.score_key, passed, detail)
-    return tool.score_frame(frame, prev_frame, region, params)
-
-
-def _multitool_has_offset(steps: list[dict[str, Any]]) -> bool:
-    """True when any chained step (idx > 0) declares an ``offset`` window.
-
-    Offset chains run the two-phase scan path and cannot resume incrementally
-    (the join needs every frame from the original start), so the worker uses
-    this to decide whether a paused task must restart from scratch.
-    """
-    return any(isinstance(s.get("offset"), dict) for s in steps[1:])
-
-
-def scan_multitool(
-    video_path: str,
-    region: dict[str, int],
-    steps: list[dict[str, Any]],
-    *,
-    start_seconds: float = 0.0,
-    end_seconds: float | None = None,
-    on_progress: Callable[[float], None] | None = None,
-    cancel_flag: Callable[[], bool] | None = None,
-    on_result: Callable[[dict[str, Any]], None] | None = None,
-    fast_opts: dict[str, Any] | None = None,
-) -> list[dict[str, Any]]:
-    """Run a multi-factor scan chaining several tool types.
-
-    Iterates the video once, checking all steps per frame.  A frame
-    must pass every step (in order) to be included in the results.
-    Results are emitted incrementally via *on_result* as each passing
-    frame is found.
-
-    Each entry in *steps* is a dict with ``"type"`` plus the tool's
-    own parameters (e.g. ``target_color``, ``tolerance`` for color).
-
-    When any step (idx > 0) carries an ``"offset"`` window
-    (``{"min", "max"}`` seconds relative to the previous step's matched
-    frame), the scan switches to a two-phase *collect-then-join* path:
-    phase 1 decodes the video once and records every step's pass/fail per
-    frame; phase 2 joins them temporally (see
-    :func:`_join_multitool_offsets`). Offset events are anchored on the
-    first step's frame and emitted after the join (not per-frame). When no
-    step has an offset the original single-pass short-circuit path runs,
-    behaving exactly as before.
-
-    Returns a list of ``{timestamp, tool_types, steps, min_confidence}``
-    dicts.
-    """
-    if len(steps) < 2:
-        raise ValueError("Multitool requires at least 2 steps")
-
-    interval = steps[0].get("interval", config.SCREENSPACE_DEFAULT_INTERVAL)
-
-    # Compute iteration range (intersection of all per-step ranges)
-    scan_start = start_seconds
-    scan_end = end_seconds
-    for step in steps:
-        s_start = step.get("start_seconds")
-        if s_start is not None:
-            scan_start = max(scan_start, s_start)
-        s_end = step.get("end_seconds")
-        if s_end is not None:
-            scan_end = min(scan_end, s_end) if scan_end is not None else s_end
-
-    vid_fps, vid_duration = _probe_video_meta(video_path)
-    if vid_fps <= 0:
-        return []
-
-    if scan_end is None or scan_end > vid_duration:
-        scan_end = vid_duration
-    total_range = scan_end - scan_start
-
-    tool_types = [s["type"] for s in steps]
-    step_regions = [s.get("region_coords", region) for s in steps]
-    prev_frame: list[np.ndarray | None] = [None]
-    results: list[dict[str, Any]] = []
-
-    def _cancel() -> bool:
-        return bool(cancel_flag and cancel_flag())
-
-    if _multitool_has_offset(steps):
-        # ---- Offset path: two-phase collect-then-join --------------------
-        # Phase 1: one decode pass, evaluate EVERY step on EVERY frame (no
-        # short-circuit), recording pass/fail + detail per sampled timestamp.
-        # ``prev_frame`` is maintained exactly as the fast path so temporal
-        # tools (change/flow/inactivity) still see the prior frame.
-        ts_list: list[float] = []
-        passed_cols: list[list[bool]] = [[] for _ in steps]
-        detail_cols: list[list[dict[str, Any] | None]] = [[] for _ in steps]
-
-        def _collect(ts: float, frame: np.ndarray) -> bool | None:
-            if _cancel():
-                return False
-            ts_list.append(round(ts, 2))
-            for i, step in enumerate(steps):
-                passed, rd = check_frame_for_tool(
-                    frame, prev_frame[0], step_regions[i], step["type"], step
-                )
-                passed_cols[i].append(bool(passed))
-                detail_cols[i].append(rd)
-            prev_frame[0] = frame
-            if on_progress and total_range > 0:
-                # Reserve the last 10% for the join (cheap, but keeps the bar
-                # from sitting at 100% while the join runs).
-                on_progress(0.9 * (ts - scan_start) / total_range)
-            return None
-
-        scan_video_full_frames(
-            video_path,
-            interval,
-            _collect,
-            start_seconds=scan_start,
-            end_seconds=scan_end,
-            fps=vid_fps,
-            duration=vid_duration,
-            fast_opts=fast_opts,
-        )
-
-        # A user cancel/pause stops the decode early; skip the join+emit and
-        # let the worker settle the cancelled/paused status. ``detect_first``
-        # does NOT trip _cancel here — it only fires on the first on_result.
-        # The join needs every frame from ``scan_start``, so this path cannot
-        # resume incrementally; ``ScreenspaceWorker.resume`` restarts offset
-        # tasks from scratch (it does not advance ``start_seconds`` for them).
-        if _cancel():
-            return []
-
-        # Phase 2: temporal join.
-        joined = _join_multitool_offsets(
-            steps, tool_types, ts_list, passed_cols, detail_cols, interval
-        )
-        emitted: list[dict[str, Any]] = []
-        for rd in joined:
-            emitted.append(rd)
-            if on_result:
-                on_result(rd)
-            if _cancel():  # detect_first stops after the first hit
-                break
-        if on_progress:
-            on_progress(1.0)
-        return emitted
-
-    def _cb(ts: float, frame: np.ndarray) -> bool | None:
-        if _cancel():
-            return False
-
-        step_results: list[dict[str, Any]] = []
-        chain_ok = True
-        for i, step in enumerate(steps):
-            passed, rd = check_frame_for_tool(
-                frame, prev_frame[0], step_regions[i], step["type"], step
-            )
-            logic = (step.get("logic") or "AND").upper() if i > 0 else "AND"
-            if logic == "NOT":
-                if passed:
-                    chain_ok = False
-                    break
-                step_results.append({"negated": True, "type": step["type"]})
-            else:
-                if not passed or rd is None:
-                    chain_ok = False
-                    break
-                step_results.append(rd)
-
-        prev_frame[0] = frame
-
-        if chain_ok and len(step_results) == len(steps):
-            confidences = []
-            for i, sr in enumerate(step_results):
-                logic = (steps[i].get("logic") or "AND").upper() if i > 0 else "AND"
-                if logic == "NOT":
-                    continue
-                confidences.append(_extract_confidence(steps[i]["type"], sr))
-            rd = {
-                "timestamp": round(ts, 2),
-                "tool_types": tool_types,
-                "steps": step_results,
-                "min_confidence": round(min(confidences), 4) if confidences else 1.0,
-            }
-            results.append(rd)
-            if on_result:
-                on_result(rd)
-
-        if on_progress and total_range > 0:
-            on_progress((ts - scan_start) / total_range)
-        return None
-
-    scan_video_full_frames(
-        video_path,
-        interval,
-        _cb,
-        start_seconds=scan_start,
-        end_seconds=scan_end,
-        fps=vid_fps,
-        duration=vid_duration,
-        fast_opts=fast_opts,
-    )
-
-    if on_progress:
-        on_progress(1.0)
-    return results
-
-
-def _join_multitool_offsets(
-    steps: list[dict[str, Any]],
-    tool_types: list[str],
-    ts_list: list[float],
-    passed_cols: list[list[bool]],
-    detail_cols: list[list[dict[str, Any] | None]],
-    interval: float,
-) -> list[dict[str, Any]]:
-    """Join per-step frame matches into offset-aware multitool events.
-
-    Phase 2 of :func:`scan_multitool`'s offset path. ``ts_list`` is the sorted
-    list of sampled timestamps; ``passed_cols[i]`` / ``detail_cols[i]`` are the
-    parallel per-step pass/fail and detail columns from phase 1.
-
-    Semantics (all confirmed with the product owner):
-
-    * **Anchor** — every step-0 match seeds a candidate chain; the emitted
-      event's ``timestamp`` is the step-0 (trigger) frame.
-    * **Cumulative** — each step ``i`` is evaluated relative to ``ref_ts``, the
-      *previous* step's matched frame. An AND match advances ``ref_ts`` to its
-      earliest in-window frame; NOT and same-frame steps leave it unchanged.
-      The advance is greedy (earliest match wins) with no backtracking — for a
-      3+ step chain a later in-window match that would have let a downstream
-      step resolve is not reconsidered.
-    * **Offset window** — ``[ref_ts + min, ref_ts + max]`` (either bound may be
-      negative). Absent offset ⇒ the exact ``ref_ts`` frame (legacy behavior).
-    * **NOT** — passes iff the condition matches in *no* frame of the window.
-
-    Adjacent anchors that resolve are coalesced (see
-    :func:`_coalesce_offset_events`).
-    """
-    n = len(steps)
-    eps = 1e-6
-    ts_to_idx = {ts: idx for idx, ts in enumerate(ts_list)}
-    logics = [
-        (steps[i].get("logic") or "AND").upper() if i > 0 else "AND" for i in range(n)
-    ]
-    offsets: list[dict[str, Any] | None] = [
-        steps[i].get("offset") if i > 0 else None for i in range(n)
-    ]
-
-    raw_events: list[dict[str, Any]] = []
-    for a, anchor_ts in enumerate(ts_list):
-        # Step 0 is always an AND anchor: needs a pass with a usable detail.
-        anchor_detail = detail_cols[0][a]
-        if not passed_cols[0][a] or anchor_detail is None:
-            continue
-        ref_ts = anchor_ts
-        chain_details: list[dict[str, Any]] = [anchor_detail]
-        chain_ok = True
-        for i in range(1, n):
-            off = offsets[i]
-            logic = logics[i]
-            if off is None:
-                # Same-frame as the previous matched frame.
-                idx = ts_to_idx.get(ref_ts)
-                hit = idx is not None and passed_cols[i][idx]
-                if logic == "NOT":
-                    if hit:
-                        chain_ok = False
-                        break
-                    chain_details.append({"negated": True, "type": steps[i]["type"]})
-                else:
-                    detail = detail_cols[i][idx] if idx is not None else None
-                    if not hit or detail is None:
-                        chain_ok = False
-                        break
-                    chain_details.append(detail)
-                    # ref_ts unchanged (same frame).
-            else:
-                lo = ref_ts + off["min"]
-                hi = ref_ts + off["max"]
-                lo_idx = bisect.bisect_left(ts_list, lo - eps)
-                hi_idx = bisect.bisect_right(ts_list, hi + eps)
-                if logic == "NOT":
-                    if any(passed_cols[i][j] for j in range(lo_idx, hi_idx)):
-                        chain_ok = False
-                        break
-                    chain_details.append({"negated": True, "type": steps[i]["type"]})
-                    # ref_ts unchanged (NOT does not advance the cursor).
-                else:
-                    match_detail: dict[str, Any] | None = None
-                    match_ts = ref_ts
-                    for j in range(lo_idx, hi_idx):
-                        if passed_cols[i][j] and detail_cols[i][j] is not None:
-                            match_detail = detail_cols[i][j]
-                            match_ts = ts_list[j]
-                            break
-                    if match_detail is None:
-                        chain_ok = False
-                        break
-                    chain_details.append(match_detail)
-                    ref_ts = match_ts  # cumulative advance to earliest match.
-        if not chain_ok:
-            continue
-        confidences = [
-            _extract_confidence(steps[i]["type"], chain_details[i])
-            for i in range(n)
-            if logics[i] != "NOT"
-        ]
-        raw_events.append(
-            {
-                "timestamp": round(anchor_ts, 2),
-                "tool_types": tool_types,
-                "steps": chain_details,
-                "min_confidence": round(min(confidences), 4) if confidences else 1.0,
-            }
-        )
-
-    return _coalesce_offset_events(raw_events, interval)
-
-
-def _coalesce_offset_events(
-    events: list[dict[str, Any]], interval: float
-) -> list[dict[str, Any]]:
-    """Collapse runs of adjacent anchor events into one representative each.
-
-    ``events`` arrive in ascending-timestamp order (anchors iterate the frame
-    grid in order). Consecutive events within ``interval * 1.5`` of one another
-    form a run; the highest-``min_confidence`` member represents the run. This
-    mirrors :func:`_merge_timestamp_spans`' adjacency rule so an offset chain
-    that fires across a burst of neighbouring frames yields a single event,
-    matching how the per-frame path's neighbours collapse downstream.
-    """
-    if not events:
-        return []
-    gap = interval * 1.5
-    coalesced: list[dict[str, Any]] = []
-    group = [events[0]]
-    for ev in events[1:]:
-        if ev["timestamp"] - group[-1]["timestamp"] <= gap:
-            group.append(ev)
-        else:
-            coalesced.append(max(group, key=lambda e: e["min_confidence"]))
-            group = [ev]
-    coalesced.append(max(group, key=lambda e: e["min_confidence"]))
-    return coalesced
-
-
-def score_multitool_frame(
-    frame: np.ndarray,
-    prev_frame: np.ndarray | None,
-    steps: list[dict[str, Any]],
-    *,
-    ocr_reader: "Callable[[str, dict[str, int], dict[str, Any]], list[Any]] | None" = None,
-) -> dict[str, Any]:
-    """Score every multitool step against one frame for pin calibration.
-
-    Unlike :func:`scan_multitool` (which short-circuits the AND chain), this
-    evaluates *all* steps unconditionally so every step row gets a score. Each
-    step uses its own ``region_coords`` (populated by the server before calling)
-    and the step dict as parameters.
-
-    Returns ``{"steps": [...], "passed": bool | None}``. The chain passes when
-    all AND steps pass and all NOT steps fail; it is ``False`` when any step
-    definitively fails; ``None`` (not evaluable) when no step definitively fails
-    but at least one could not be scored.
-
-    Calibration is single-frame, so a step's ``offset`` window is **not**
-    evaluated here — each step is scored against this one pinned frame. Offset
-    chains are only resolved temporally at scan time (:func:`scan_multitool`).
-    """
-    step_results: list[dict[str, Any]] = []
-    definitive_fail = False
-    any_not_evaluable = False
-    for i, step in enumerate(steps):
-        region = step.get("region_coords", {})
-        res = score_frame_for_tool(
-            step["type"], frame, prev_frame, region, step, ocr_reader=ocr_reader
-        )
-        logic = (step.get("logic") or "AND").upper() if i > 0 else "AND"
-        entry = dict(res)
-        entry["type"] = step["type"]
-        entry["logic"] = logic
-        step_results.append(entry)
-        if res.get("status") != "ok":
-            any_not_evaluable = True
-            continue
-        step_passed = bool(res.get("passed"))
-        if logic == "NOT":
-            if step_passed:
-                definitive_fail = True
-        elif not step_passed:
-            definitive_fail = True
-    if definitive_fail:
-        chain_passed: bool | None = False
-    elif any_not_evaluable:
-        chain_passed = None
-    else:
-        chain_passed = True
-    return {"steps": step_results, "passed": chain_passed}
-
-
-def _merge_timestamp_spans(
-    timestamps: list[float], interval: float
-) -> list[dict[str, Any]]:
-    """Merge consecutive matched timestamps into spans."""
-    if not timestamps:
-        return []
-    timestamps.sort()
-    spans: list[dict[str, Any]] = []
-    start = timestamps[0]
-    end = timestamps[0]
-    gap = interval * 1.5
-
-    for ts in timestamps[1:]:
-        if ts - end <= gap:
-            end = ts
-        else:
-            spans.append(
-                {
-                    "start": round(start, 2),
-                    "end": round(end, 2),
-                    "duration": round(end - start, 2),
-                }
-            )
-            start = ts
-            end = ts
-
-    spans.append(
-        {
-            "start": round(start, 2),
-            "end": round(end, 2),
-            "duration": round(end - start, 2),
-        }
-    )
-    return spans
-
-
-# ---------------------------------------------------------------------------
-# Task queue and worker
-# ---------------------------------------------------------------------------
-
-TASK_STATUS_QUEUED = "queued"
-TASK_STATUS_RUNNING = "running"
-TASK_STATUS_COMPLETED = "completed"
-TASK_STATUS_FAILED = "failed"
-TASK_STATUS_CANCELLED = "cancelled"
-TASK_STATUS_PAUSED = "paused"
-
-_SENTINEL = object()
-
-
-def create_task(
-    task_type: str,
-    participant: str,
-    source_video: str,
-    video_paths: list[str],
-    region_name: str,
-    region_coords: dict[str, int],
-    parameters: dict[str, Any] | None = None,
-) -> dict[str, Any]:
-    """Build a new task dict with all fields initialized.
-
-    *video_paths* is the participant's ordered source video(s): one entry for a
-    normal participant, several for a multi-video participant whose session spans
-    files (analyzed as one continuous timeline; the worker maps the task's global
-    time range into the owning sub-video per part).
-    """
-    return {
-        "id": f"ss_{uuid.uuid4().hex[:8]}",
-        "type": task_type,
-        "participant": participant,
-        "source_video": source_video,
-        "video_paths": video_paths,
-        "region": region_name,
-        "region_coords": region_coords,
-        "parameters": parameters or {},
-        "status": TASK_STATUS_QUEUED,
-        "progress": 0.0,
-        "priority": 100,
-        "result": None,
-        "error": None,
-        "created_at": datetime.now(timezone.utc).isoformat(),
-        "completed_at": None,
-        "_cancelled": False,
-    }
-
-
-# ---------------------------------------------------------------------------
-# Heatmap generation
-# ---------------------------------------------------------------------------
-
-
-def generate_template_heatmap(
-    results: list[dict[str, Any]],
-    frame_width: int,
-    frame_height: int,
-    output_path: str,
-) -> str | None:
-    """Generate a heatmap PNG from accumulated template match bounding boxes.
-
-    Each pixel's intensity reflects how many times (weighted by score) it fell
-    inside a match bounding box across all scanned frames.
-    """
-    accumulator = np.zeros((frame_height, frame_width), dtype=np.float32)
-    for r in results:
-        for m in r.get("matches", []):
-            x, y, w, h = int(m["x"]), int(m["y"]), int(m["w"]), int(m["h"])
-            y2 = min(y + h, frame_height)
-            x2 = min(x + w, frame_width)
-            accumulator[y:y2, x:x2] += m.get("score", 1.0)
-
-    if accumulator.max() == 0:
-        return None
-
-    normalized = (accumulator / accumulator.max() * 255).astype(np.uint8)
-    normalized = cv2.GaussianBlur(normalized, (15, 15), 0)
-    heatmap = cv2.applyColorMap(normalized, cv2.COLORMAP_JET)
-    cv2.imwrite(output_path, heatmap)
-    return output_path
-
-
-def generate_flow_heatmap(
-    results: list[dict[str, Any]],
-    region_width: int,
-    region_height: int,
-    output_path: str,
-) -> str | None:
-    """Generate a heatmap PNG from accumulated optical flow magnitudes.
-
-    Uses ``flow_grid`` data from each result to paint per-cell motion
-    intensity across all frames.
-    """
-    acc_size = 256
-    accumulator = np.zeros((acc_size, acc_size), dtype=np.float32)
-    for r in results:
-        for cell in r.get("flow_grid", []):
-            cx = int(cell["x"] * (acc_size - 1))
-            cy = int(cell["y"] * (acc_size - 1))
-            radius = max(1, acc_size // 16)
-            cv2.circle(accumulator, (cx, cy), radius, float(cell["mag"]), -1)
-
-    if accumulator.max() == 0:
-        return None
-
-    normalized = (accumulator / accumulator.max() * 255).astype(np.uint8)
-    normalized = cv2.GaussianBlur(normalized, (15, 15), 0)
-    heatmap = cv2.applyColorMap(normalized, cv2.COLORMAP_JET)
-    heatmap = cv2.resize(
-        heatmap, (region_width, region_height), interpolation=cv2.INTER_LINEAR
-    )
-    cv2.imwrite(output_path, heatmap)
-    return output_path
-
-
-def _accumulate_heatmap_result(
-    accumulator: np.ndarray,
-    result: dict[str, Any],
-    heatmap_type: str,
-) -> None:
-    """Add a single result's contribution to a heatmap accumulator."""
-    acc_h, acc_w = accumulator.shape[:2]
-    if heatmap_type == "template":
-        for m in result.get("matches", []):
-            x, y, w, h = int(m["x"]), int(m["y"]), int(m["w"]), int(m["h"])
-            y2 = min(y + h, acc_h)
-            x2 = min(x + w, acc_w)
-            accumulator[y:y2, x:x2] += m.get("score", 1.0)
-    elif heatmap_type == "flow":
-        for cell in result.get("flow_grid", []):
-            cx = int(cell["x"] * (acc_w - 1))
-            cy = int(cell["y"] * (acc_h - 1))
-            radius = max(1, acc_w // 16)
-            cv2.circle(accumulator, (cx, cy), radius, float(cell["mag"]), -1)
-
-
-def generate_heatmap_gif(
-    results: list[dict[str, Any]],
-    width: int,
-    height: int,
-    output_path: str,
-    heatmap_type: str = "template",
-    num_frames: int = 24,
-    frame_duration_ms: int = 120,
-) -> str | None:
-    """Generate an animated GIF showing heatmap accumulation over time.
-
-    Divides *results* into *num_frames* temporal buckets, progressively
-    accumulates heatmap data, and writes frames as an animated GIF.
-    """
-    from PIL import Image
-
-    if not results:
-        return None
-
-    num_frames = min(num_frames, len(results))
-    if num_frames < 2:
-        return None
-
-    acc_h, acc_w = height, width
-    if heatmap_type == "flow":
-        acc_h = acc_w = 256
-
-    # First pass: compute global max for consistent normalization
-    global_acc = np.zeros((acc_h, acc_w), dtype=np.float32)
-    for r in results:
-        _accumulate_heatmap_result(global_acc, r, heatmap_type)
-    global_max = float(global_acc.max())
-    if global_max == 0:
-        return None
-
-    # Second pass: build progressive frames
-    accumulator = np.zeros((acc_h, acc_w), dtype=np.float32)
-    frames: list[Image.Image] = []
-    bucket_size = max(1, len(results) // num_frames)
-
-    for frame_idx in range(num_frames):
-        start_idx = frame_idx * bucket_size
-        end_idx = (
-            len(results)
-            if frame_idx == num_frames - 1
-            else min((frame_idx + 1) * bucket_size, len(results))
-        )
-        for r_idx in range(start_idx, end_idx):
-            _accumulate_heatmap_result(accumulator, results[r_idx], heatmap_type)
-
-        normalized = (accumulator / global_max * 255).astype(np.uint8)
-        normalized = cv2.GaussianBlur(normalized, (15, 15), 0)
-        colored = cv2.applyColorMap(normalized, cv2.COLORMAP_JET)
-
-        if heatmap_type == "flow":
-            colored = cv2.resize(
-                colored, (width, height), interpolation=cv2.INTER_LINEAR
-            )
-
-        rgb = cv2.cvtColor(colored, cv2.COLOR_BGR2RGB)
-        frames.append(Image.fromarray(rgb))
-
-    if not frames:
-        return None
-
-    frames[0].save(
-        output_path,
-        save_all=True,
-        append_images=frames[1:],
-        duration=frame_duration_ms,
-        loop=0,
-    )
-    return output_path
-
-
-# ---------------------------------------------------------------------------
-# Analysis tools (strategy registry)
-# ---------------------------------------------------------------------------
-#
-# Each tool is a small class wrapping the corresponding module-level ``scan_*``
-# function (preserved for tests that monkeypatch them). The two registry-level
-# dispatch points are:
-#   - :func:`check_frame_for_tool` (single-frame eval used by multitool)
-#   - :meth:`ScreenspaceWorker._dispatch` (full-video scan, called by the worker)
-# Both look up the tool by name in ``TOOLS`` and delegate to its methods.
-
-
-def _numbers_ocr_allowlist(languages: list[str], integers_only: bool) -> str | None:
-    """EasyOCR allowlist for the numbers tool (English only; ``None`` otherwise)."""
-    if languages == ["en"]:
-        return _OCR_DIGITS_ONLY_ALLOWLIST if integers_only else _OCR_NUMBER_ALLOWLIST
-    return None
-
-
-def _ocr_region_readings(
-    region_pixels: np.ndarray,
-    *,
-    languages: list[str] | None = None,
-    allowlist: str | None = None,
-    preprocess: bool = False,
-) -> list[Any]:
-    """Run EasyOCR over a region crop and return raw ``(bbox, text, conf)`` tuples.
-
-    Pure transport over the cached reader — no fuzzy/threshold logic — so the
-    same readings can be re-scored under different fuzzy/confidence settings
-    (the basis of the calibration OCR cache).
-    """
-    langs = languages or ["en"]
-    reader = _get_ocr_reader(langs)
-    pixels = _preprocess_for_ocr(region_pixels) if preprocess else region_pixels
-    kwargs: dict[str, Any] = {"detail": 1}
-    if allowlist is not None:
-        kwargs["allowlist"] = allowlist
-    return reader.readtext(pixels, **kwargs)
-
-
-def _score_text_readings(
-    readings: list[Any], params: dict[str, Any]
-) -> tuple[bool, dict[str, Any]]:
-    """Score text OCR readings: best fuzzy ratio among readings clearing min conf.
-
-    ``fuzzy_ratio`` is the calibration scalar; ``text_found``/``confidence`` carry
-    the best-matching reading for the strip tooltip. ``passed`` is the fuzzy
-    match at the current threshold.
-    """
-    search_string = params.get("search_string", "")
-    fuzzy_threshold = params.get(
-        "fuzzy_threshold", config.SCREENSPACE_OCR_FUZZY_THRESHOLD
-    )
-    ocr_min_conf = _effective_ocr_confidence_threshold(
-        params.get("ocr_confidence_threshold")
-    )
-    ocr_normalize = params.get("ocr_normalize") or "off"
-    if ocr_normalize not in ("digits", "letters"):
-        ocr_normalize = "off"
-    search_cmp = _normalize_ocr_text(search_string, ocr_normalize)
-    best_ratio = 0.0
-    best_text = ""
-    best_conf = 0.0
-    for _, text, conf in readings:
-        if conf < ocr_min_conf:
-            continue
-        ocr_cmp = _normalize_ocr_text(text, ocr_normalize)
-        ratio = difflib.SequenceMatcher(None, search_cmp, ocr_cmp).ratio()
-        if ratio > best_ratio:
-            best_ratio = ratio
-            best_text = text
-            best_conf = conf
-    return best_ratio >= fuzzy_threshold, {
-        "fuzzy_ratio": round(best_ratio, 4),
-        "text_found": best_text,
-        "confidence": round(best_conf, 4),
-    }
-
-
-def _score_numbers_readings(
-    readings: list[Any], params: dict[str, Any]
-) -> tuple[bool, dict[str, Any]]:
-    """Score numbers OCR readings by the best condition-matching OCR confidence.
-
-    ``confidence`` is the calibration scalar for the OCR-confidence slider: it
-    reflects the best reading that satisfies operator / target_value / range,
-    not an unrelated high-confidence number. ``passed`` then applies the current
-    confidence threshold to that matching reading.
-    """
-    operator = params.get("operator", "gt")
-    target_value = params.get("target_value", 0)
-    range_min = params.get("range_min")
-    range_max = params.get("range_max")
-    ocr_min_conf = _effective_ocr_confidence_threshold(
-        params.get("ocr_confidence_threshold")
-    )
-    matched_number: float | None = None
-    matched_conf = 0.0
-    for _, text, conf in readings:
-        cleaned = text.replace(",", "")
-        for match in _NUMBERS_RE.findall(cleaned):
-            num = float(match)
-            if _number_matches(num, operator, target_value, range_min, range_max):
-                if matched_number is None or conf > matched_conf:
-                    matched_number = num
-                    matched_conf = conf
-    detail: dict[str, Any] = {"confidence": round(matched_conf, 4)}
-    if matched_number is not None:
-        detail["number_found"] = matched_number
-    return matched_number is not None and matched_conf >= ocr_min_conf, detail
-
-
-def run_calibration_ocr(
-    frame: np.ndarray,
-    region: dict[str, int],
-    tool_type: str,
-    params: dict[str, Any],
-) -> list[Any]:
-    """Run EasyOCR for one calibration frame/region (text or numbers tool).
-
-    Public entry point for the server's per-pin OCR cache: returns the raw
-    ``(bbox, text, conf)`` readings so they can be memoized and re-scored under
-    changed fuzzy/confidence settings without re-running OCR.
-    """
-    languages = params.get("languages") or ["en"]
-    allowlist = (
-        _numbers_ocr_allowlist(languages, params.get("integers_only", False))
-        if tool_type == "numbers"
-        else None
-    )
-    pixels = extract_region(frame, region)
-    return _ocr_region_readings(
-        pixels,
-        languages=languages,
-        allowlist=allowlist,
-        preprocess=params.get("ocr_preprocess", False),
-    )
-
-
-class AnalysisTool:
-    """Base class for screenspace analysis tools.
-
-    Subclasses set ``name`` and override ``check_frame`` (for multitool
-    chaining) and/or ``scan`` (for the full-video sweep). The dispatch
-    layer reads ``fast_scan_region_dim`` / ``supports_fast_scan`` /
-    ``fast_scan_extra_opts`` to build the shared ``fast_opts`` payload.
-    """
-
-    name: ClassVar[str] = ""
-    # Max region dimension when running in "fast" scan mode (passed to the
-    # generic frame extractor as ``max_region_dim``). 0 means no downscale.
-    fast_scan_region_dim: ClassVar[int] = 0
-    # Whether the tool participates in the fast-scan optimization at all.
-    # Timelapse opts out because it has its own ``sample_interval``.
-    supports_fast_scan: ClassVar[bool] = True
-    # Extra keys merged into ``fast_opts`` (e.g. ``{"template_downscale": True}``).
-    fast_scan_extra_opts: ClassVar[dict[str, Any]] = {}
-    # Detail-dict key holding the threshold-independent calibration scalar that
-    # ``check_frame`` populates on both branches. Empty ⇒ tool not calibratable.
-    score_key: ClassVar[str] = ""
-
-    def check_frame(
-        self,
-        frame: np.ndarray,
-        prev_frame: np.ndarray | None,
-        region: dict[str, int],
-        params: dict[str, Any],
-    ) -> tuple[bool, dict[str, Any] | None]:
-        """Evaluate a single frame for use in a multitool chain step.
-
-        Default: tool not supported as a multitool step.
-        """
-        return False, None
-
-    def score_frame(
-        self,
-        frame: np.ndarray,
-        prev_frame: np.ndarray | None,
-        region: dict[str, int],
-        params: dict[str, Any],
-    ) -> dict[str, Any]:
-        """Return the threshold-independent calibration score for one frame.
-
-        Reads the scalar ``check_frame`` now populates on both branches (keyed by
-        ``score_key``). Returns ``{"status": "not_evaluable"}`` when the frame
-        cannot be scored (missing companion/reference, degenerate input, etc.).
-        """
-        passed, detail = self.check_frame(frame, prev_frame, region, params)
-        return _score_result(self.score_key, passed, detail)
-
-    def scan(
-        self,
-        video_path: str,
-        region: dict[str, int],
-        params: dict[str, Any],
-        *,
-        task_id: str,
-        scan_mode: str,
-        on_progress: Callable[[float], None],
-        cancel_flag: Callable[[], bool],
-        on_result: Callable[[dict[str, Any]], None] | None,
-        fast_opts: dict[str, Any] | None,
-    ) -> Any:
-        """Run a full-video scan. Subclasses must override."""
-        raise NotImplementedError
-
-
-class ColorTool(AnalysisTool):
-    name = "color"
-    fast_scan_region_dim = 32
-    score_key = "_confidence"
-
-    def check_frame(self, frame, prev_frame, region, params):
-        pixels = extract_region(frame, region)
-        target = params.get("target_color", {"h": 0, "s": 0, "v": 0})
-        tol = params.get("tolerance", {"h": 10, "s": 50, "v": 50})
-        if params.get("color_mode") == "presence":
-            matched, conf = color_present(
-                pixels, target, tol, params.get("min_coverage", 0.0)
-            )
-        else:
-            matched, conf = color_matches(pixels, target, tol)
-        return matched, {"_confidence": conf}
-
-    def scan(
-        self,
-        video_path,
-        region,
-        params,
-        *,
-        task_id,
-        scan_mode,
-        on_progress,
-        cancel_flag,
-        on_result,
-        fast_opts,
-    ):
-        color_mode = params.get("color_mode", "average")
-        # Presence scans must stay full-resolution: the fast-scan max_region_dim
-        # downscale uses INTER_AREA averaging, which erases small color patches.
-        if color_mode == "presence":
-            fast_opts = None
-        return scan_color(
-            video_path,
-            region,
-            target_color=params.get("target_color", {"h": 0, "s": 0, "v": 0}),
-            tolerance=params.get("tolerance", {"h": 10, "s": 50, "v": 50}),
-            interval_seconds=params.get("interval", 0),
-            start_seconds=params.get("start_seconds", 0.0),
-            end_seconds=params.get("end_seconds"),
-            color_mode=color_mode,
-            min_coverage=params.get("min_coverage", 0.0),
-            on_progress=on_progress,
-            cancel_flag=cancel_flag,
-            on_result=on_result,
-            fast_opts=fast_opts,
-        )
-
-
-class ChangeTool(AnalysisTool):
-    name = "change"
-    fast_scan_region_dim = 128
-    score_key = "magnitude"
-
-    def check_frame(self, frame, prev_frame, region, params):
-        if prev_frame is None:
-            return False, None
-        pixels = extract_region(frame, region)
-        prev_pixels = extract_region(prev_frame, region)
-        threshold = params.get("threshold", config.SCREENSPACE_CHANGE_RATIO_THRESHOLD)
-        noise_threshold = params.get(
-            "noise_threshold", config.SCREENSPACE_NOISE_THRESHOLD
-        )
-        k = config.SCREENSPACE_BLUR_KERNEL
-        morph_kernel = _morph_kernel(config.SCREENSPACE_MORPH_KERNEL)
-        curr_gray = cv2.cvtColor(
-            cv2.GaussianBlur(pixels, (k, k), 0), cv2.COLOR_BGR2GRAY
-        )
-        prev_gray = cv2.cvtColor(
-            cv2.GaussianBlur(prev_pixels, (k, k), 0), cv2.COLOR_BGR2GRAY
-        )
-        diff = cv2.absdiff(prev_gray, curr_gray)
-        _, mask = cv2.threshold(diff, noise_threshold, 255, cv2.THRESH_BINARY)
-        mask = cv2.morphologyEx(mask, cv2.MORPH_OPEN, morph_kernel)
-        mag = float(np.count_nonzero(mask)) / float(mask.size) if mask.size else 0.0
-        return mag >= threshold, {"magnitude": round(mag, 4)}
-
-    def scan(
-        self,
-        video_path,
-        region,
-        params,
-        *,
-        task_id,
-        scan_mode,
-        on_progress,
-        cancel_flag,
-        on_result,
-        fast_opts,
-    ):
-        return scan_changes(
-            video_path,
-            region,
-            threshold=params.get("threshold", 0),
-            interval_seconds=params.get("interval", 0),
-            noise_threshold=params.get("noise_threshold", 0),
-            require_consecutive=params.get("require_consecutive", 1),
-            start_seconds=params.get("start_seconds", 0.0),
-            end_seconds=params.get("end_seconds"),
-            on_progress=on_progress,
-            cancel_flag=cancel_flag,
-            on_result=on_result,
-            fast_opts=fast_opts,
-        )
-
-
-class SimilarityTool(AnalysisTool):
-    name = "similarity"
-    fast_scan_region_dim = 128
-    score_key = "score"
-
-    def check_frame(self, frame, prev_frame, region, params):
-        ref = params.get("reference_frame")
-        if ref is None:
-            return False, None
-        pixels = extract_region(frame, region)
-        threshold = params.get("threshold", config.SCREENSPACE_SSIM_THRESHOLD)
-        is_sim, score = regions_are_similar(pixels, ref, threshold)
-        return is_sim, {"score": round(score, 4)}
-
-    def scan(
-        self,
-        video_path,
-        region,
-        params,
-        *,
-        task_id,
-        scan_mode,
-        on_progress,
-        cancel_flag,
-        on_result,
-        fast_opts,
-    ):
-        ref_frame = params.get("reference_frame")
-        if ref_frame is None:
-            raise ValueError("Similarity scan requires a reference_frame parameter")
-        return scan_similarity(
-            video_path,
-            region,
-            reference_frame=ref_frame,
-            threshold=params.get("threshold", 0),
-            interval_seconds=params.get("interval", 0),
-            start_seconds=params.get("start_seconds", 0.0),
-            end_seconds=params.get("end_seconds"),
-            on_progress=on_progress,
-            cancel_flag=cancel_flag,
-            on_result=on_result,
-            fast_opts=fast_opts,
-        )
-
-
-class TextTool(AnalysisTool):
-    name = "text"
-    # Calibration scalar is fuzzy match quality. Deliberately distinct from
-    # ``_extract_confidence``'s "confidence" key (OCR reading confidence) — two
-    # different axes; do not unify them.
-    score_key = "fuzzy_ratio"
-
-    def check_frame(self, frame, prev_frame, region, params):
-        if not params.get("search_string", ""):
-            return False, None
-        pixels = extract_region(frame, region)
-        readings = _ocr_region_readings(
-            pixels,
-            languages=params.get("languages") or ["en"],
-            preprocess=params.get("ocr_preprocess", False),
-        )
-        return _score_text_readings(readings, params)
-
-    def scan(
-        self,
-        video_path,
-        region,
-        params,
-        *,
-        task_id,
-        scan_mode,
-        on_progress,
-        cancel_flag,
-        on_result,
-        fast_opts,
-    ):
-        return scan_text(
-            video_path,
-            region,
-            search_string=params.get("search_string", ""),
-            interval_seconds=params.get("interval", 2.0),
-            fuzzy_threshold=params.get("fuzzy_threshold", 0),
-            ocr_confidence_threshold=params.get("ocr_confidence_threshold"),
-            ocr_preprocess=params.get("ocr_preprocess", False),
-            ocr_normalize=params.get("ocr_normalize") or "off",
-            require_consecutive=params.get("require_consecutive", 1),
-            languages=params.get("languages"),
-            start_seconds=params.get("start_seconds", 0.0),
-            end_seconds=params.get("end_seconds"),
-            on_progress=on_progress,
-            cancel_flag=cancel_flag,
-            on_result=on_result,
-            fast_opts=fast_opts,
-        )
-
-
-class NumbersTool(AnalysisTool):
-    name = "numbers"
-    score_key = "confidence"
-
-    def check_frame(self, frame, prev_frame, region, params):
-        languages = params.get("languages") or ["en"]
-        pixels = extract_region(frame, region)
-        readings = _ocr_region_readings(
-            pixels,
-            languages=languages,
-            allowlist=_numbers_ocr_allowlist(
-                languages, params.get("integers_only", False)
-            ),
-            preprocess=params.get("ocr_preprocess", False),
-        )
-        return _score_numbers_readings(readings, params)
-
-    def scan(
-        self,
-        video_path,
-        region,
-        params,
-        *,
-        task_id,
-        scan_mode,
-        on_progress,
-        cancel_flag,
-        on_result,
-        fast_opts,
-    ):
-        return scan_numbers(
-            video_path,
-            region,
-            operator=params.get("operator", "gt"),
-            target_value=params.get("target_value", 0),
-            interval_seconds=params.get("interval", 2.0),
-            range_min=params.get("range_min"),
-            range_max=params.get("range_max"),
-            ocr_confidence_threshold=params.get("ocr_confidence_threshold"),
-            ocr_preprocess=params.get("ocr_preprocess", False),
-            integers_only=params.get("integers_only", False),
-            require_consecutive=params.get("require_consecutive", 1),
-            languages=params.get("languages"),
-            start_seconds=params.get("start_seconds", 0.0),
-            end_seconds=params.get("end_seconds"),
-            on_progress=on_progress,
-            cancel_flag=cancel_flag,
-            on_result=on_result,
-            fast_opts=fast_opts,
-        )
-
-
-class TemplateTool(AnalysisTool):
-    name = "template"
-    fast_scan_extra_opts: ClassVar[dict[str, Any]] = {"template_downscale": True}
-    score_key = "best_score"
-
-    def check_frame(self, frame, prev_frame, region, params):
-        template_img = params.get("template_image")
-        if template_img is None:
-            return False, None
-        threshold = params.get("threshold", config.SCREENSPACE_TEMPLATE_MATCH_THRESHOLD)
-        # Cache the per-task-constant scaled template/mask + grayscale prep on the
-        # parameters dict so multitool scans amortize it across frames. The
-        # template_scale slider resizes the (often uploaded) template to its
-        # in-video pixel size before matching, mirroring scan_template.
-        cached = params.get("_prepared_template")
-        if cached is None:
-            scaled_img, scaled_mask = _scale_template(
-                template_img,
-                params.get("template_mask"),
-                float(params.get("template_scale", 1.0)),
-            )
-            cached = (
-                scaled_img,
-                scaled_mask,
-                _prepare_template(scaled_img, scaled_mask),
-            )
-            params["_prepared_template"] = cached
-        scaled_img, scaled_mask, prepared = cached
-        # Peak correlation is the threshold-independent scalar; available even on
-        # a miss (unlike match_template, which only returns above-threshold hits).
-        corr = _template_correlation_map(frame, prepared)
-        if corr is None:
-            return False, None
-        peak = float(corr.max())
-        if peak < threshold:
-            return False, {"best_score": round(peak, 4), "match_count": 0}
-        matches = match_template(
-            frame,
-            scaled_img,
-            threshold=threshold,
-            mask=scaled_mask,
-            prepared=prepared,
-        )
-        return True, {"best_score": round(peak, 4), "match_count": len(matches)}
-
-    def scan(
-        self,
-        video_path,
-        region,
-        params,
-        *,
-        task_id,
-        scan_mode,
-        on_progress,
-        cancel_flag,
-        on_result,
-        fast_opts,
-    ):
-        template_img = params.get("template_image")
-        if template_img is None:
-            raise ValueError("Template scan requires a template_image parameter")
-        tmpl_mask = params.get("template_mask")
-        # Fast scan: downscale template + mask by 2x before passing to the
-        # scan function (which separately downscales the frame via the
-        # ``template_downscale`` fast_opts flag).
-        if scan_mode == "fast":
-            th, tw = template_img.shape[:2]
-            ntw, nth = tw // 2, th // 2
-            if ntw > 0 and nth > 0:
-                template_img = cv2.resize(
-                    template_img, (ntw, nth), interpolation=cv2.INTER_AREA
-                )
-                if tmpl_mask is not None:
-                    tmpl_mask = cv2.resize(
-                        tmpl_mask, (ntw, nth), interpolation=cv2.INTER_AREA
-                    )
-        return scan_template(
-            video_path,
-            region,
-            template_image=template_img,
-            threshold=params.get("threshold", 0),
-            interval_seconds=params.get("interval", 0),
-            template_mask=tmpl_mask,
-            template_scale=float(params.get("template_scale", 1.0)),
-            start_seconds=params.get("start_seconds", 0.0),
-            end_seconds=params.get("end_seconds"),
-            on_progress=on_progress,
-            cancel_flag=cancel_flag,
-            on_result=on_result,
-            fast_opts=fast_opts,
-        )
-
-
-class FlowTool(AnalysisTool):
-    name = "flow"
-    fast_scan_region_dim = 128
-    score_key = "magnitude"
-
-    def check_frame(self, frame, prev_frame, region, params):
-        if prev_frame is None:
-            return False, None
-        pixels = extract_region(frame, region)
-        prev_pixels = extract_region(prev_frame, region)
-        curr_gray = cv2.cvtColor(pixels, cv2.COLOR_BGR2GRAY)
-        prev_gray_f = cv2.cvtColor(prev_pixels, cv2.COLOR_BGR2GRAY)
-        magnitude_threshold = params.get(
-            "magnitude_threshold", config.SCREENSPACE_FLOW_MAGNITUDE_THRESHOLD
-        )
-        flow_result = compute_optical_flow(prev_gray_f, curr_gray)
-        return flow_result["magnitude"] >= magnitude_threshold, {
-            "magnitude": flow_result["magnitude"],
-            "angle": flow_result["angle"],
-        }
-
-    def scan(
-        self,
-        video_path,
-        region,
-        params,
-        *,
-        task_id,
-        scan_mode,
-        on_progress,
-        cancel_flag,
-        on_result,
-        fast_opts,
-    ):
-        return scan_flow(
-            video_path,
-            region,
-            magnitude_threshold=params.get("magnitude_threshold", 0),
-            interval_seconds=params.get("interval", 0),
-            require_consecutive=params.get("require_consecutive", 1),
-            start_seconds=params.get("start_seconds", 0.0),
-            end_seconds=params.get("end_seconds"),
-            on_progress=on_progress,
-            cancel_flag=cancel_flag,
-            on_result=on_result,
-            fast_opts=fast_opts,
-        )
-
-
-class SceneTool(AnalysisTool):
-    name = "scene"
-    fast_scan_region_dim = 64
-    score_key = "score"
-
-    def check_frame(self, frame, prev_frame, region, params):
-        ref_scenes = params.get("reference_scenes")
-        if not ref_scenes:
-            return False, None
-        threshold = params.get(
-            "threshold", config.SCREENSPACE_SCENE_SIMILARITY_THRESHOLD
-        )
-        pixels = extract_region(frame, region)
-        fp = compute_scene_fingerprint(pixels)
-        best_name = ""
-        best_score = 0.0
-        for ref in ref_scenes:
-            # Cache fingerprint on the reference dict to avoid recomputing per frame
-            ref_fp = ref.get("_cached_fingerprint")
-            if ref_fp is None:
-                ref_fp = compute_scene_fingerprint(ref["frame"])
-                ref["_cached_fingerprint"] = ref_fp
-            score = compare_scene_fingerprints(fp, ref_fp)
-            if score > best_score:
-                best_score = score
-                best_name = ref["name"]
-        return best_score >= threshold, {
-            "scene_name": best_name,
-            "score": round(best_score, 4),
-        }
-
-    def scan(
-        self,
-        video_path,
-        region,
-        params,
-        *,
-        task_id,
-        scan_mode,
-        on_progress,
-        cancel_flag,
-        on_result,
-        fast_opts,
-    ):
-        ref_scenes = params.get("reference_scenes")
-        if not ref_scenes:
-            raise ValueError("Scene scan requires reference_scenes parameter")
-        return scan_scene(
-            video_path,
-            region,
-            reference_scenes=ref_scenes,
-            threshold=params.get("threshold", 0),
-            interval_seconds=params.get("interval", 0),
-            start_seconds=params.get("start_seconds", 0.0),
-            end_seconds=params.get("end_seconds"),
-            on_progress=on_progress,
-            cancel_flag=cancel_flag,
-            on_result=on_result,
-            fast_opts=fast_opts,
-        )
-
-
-class InactivityTool(AnalysisTool):
-    name = "inactivity"
-    fast_scan_region_dim = 64
-    # Calibration scalar is the raw phash distance (Sensitivity-slider units);
-    # the strip inverts the axis for display (lower distance = more inactive).
-    score_key = "distance"
-
-    def check_frame(self, frame, prev_frame, region, params):
-        if prev_frame is None:
-            return False, None
-        pixels = extract_region(frame, region)
-        prev_pixels = extract_region(prev_frame, region)
-        thresh = params.get("threshold", config.SCREENSPACE_INACTIVITY_PHASH_THRESHOLD)
-        curr_h = compute_phash(pixels)
-        prev_h = compute_phash(prev_pixels)
-        dist = int(curr_h - prev_h)
-        if thresh > 0:
-            conf = max(0.0, min((thresh - dist) / float(thresh), 1.0))
-        else:
-            conf = 1.0 if dist <= thresh else 0.0
-        return dist <= thresh, {"distance": dist, "_confidence": conf}
-
-    def scan(
-        self,
-        video_path,
-        region,
-        params,
-        *,
-        task_id,
-        scan_mode,
-        on_progress,
-        cancel_flag,
-        on_result,
-        fast_opts,
-    ):
-        return scan_inactivity(
-            video_path,
-            region,
-            threshold=params.get("threshold", 0),
-            min_duration=params.get("min_duration", 0.0),
-            interval_seconds=params.get("interval", 0),
-            start_seconds=params.get("start_seconds", 0.0),
-            end_seconds=params.get("end_seconds"),
-            on_progress=on_progress,
-            cancel_flag=cancel_flag,
-            on_result=on_result,
-            fast_opts=fast_opts,
-        )
-
-
-class BoundaryTool(AnalysisTool):
-    name = "boundary"
-    # The scanner is already coarse and runs its own phash on every sample;
-    # the generic fast-scan phash-skip would fight that logic, so opt out.
-    supports_fast_scan = False
-    # Scan-only in v1: not a multitool step and not calibratable (no score_key),
-    # so the pinned-frame strip and /api/calibrate correctly skip it. Wiring
-    # calibration later means adding score_key + a per-frame check_frame here,
-    # plus a `needs_prev` entry and full-frame handling in the calibrate endpoint
-    # (boundary needs the previous sampled frame and always scans the full frame).
-
-    def scan(
-        self,
-        video_path,
-        region,
-        params,
-        *,
-        task_id,
-        scan_mode,
-        on_progress,
-        cancel_flag,
-        on_result,
-        fast_opts,
-    ):
-        return scan_boundaries(
-            video_path,
-            region,
-            threshold=params.get("threshold", 0),
-            min_gap=params.get("min_gap", 0.0),
-            interval_seconds=params.get("interval", 0),
-            start_seconds=params.get("start_seconds", 0.0),
-            end_seconds=params.get("end_seconds"),
-            on_progress=on_progress,
-            cancel_flag=cancel_flag,
-            on_result=on_result,
-            fast_opts=fast_opts,
-        )
-
-
-class TimelapseTool(AnalysisTool):
-    name = "timelapse"
-    # Has its own ``sample_interval`` and produces a media file rather than
-    # per-frame events, so the generic fast-scan path does not apply.
-    supports_fast_scan = False
-
-    def scan(
-        self,
-        video_path,
-        region,
-        params,
-        *,
-        task_id,
-        scan_mode,
-        on_progress,
-        cancel_flag,
-        on_result,
-        fast_opts,
-    ):
-        output_path = params.get("output_path", "")
-        if not output_path:
-            ext = "gif" if params.get("output_format") == "gif" else "mp4"
-            output_path = str(
-                Path(utils.get_effective_output_dir()) / f"timelapse_{task_id}.{ext}"
-            )
-        return generate_timelapse(
-            video_path,
-            region,
-            speedup_factor=params.get("speedup_factor", 10.0),
-            output_path=output_path,
-            output_format=params.get("output_format", "mp4"),
-            start_seconds=params.get("start_seconds", 0.0),
-            end_seconds=params.get("end_seconds"),
-            sample_interval=params.get("sample_interval", 0.0),
-            on_progress=on_progress,
-            cancel_flag=cancel_flag,
-        )
-
-
-class MultitoolTool(AnalysisTool):
-    name = "multitool"
-
-    def scan(
-        self,
-        video_path,
-        region,
-        params,
-        *,
-        task_id,
-        scan_mode,
-        on_progress,
-        cancel_flag,
-        on_result,
-        fast_opts,
-    ):
-        steps = [dict(s) for s in params.get("steps", [])]
-        if len(steps) < 2:
-            raise ValueError("Multitool requires at least 2 steps")
-        # Fast scan: multiply the interval used by scan_multitool
-        # (reads from steps[0]["interval"] with fallback to default).
-        if scan_mode == "fast" and steps:
-            mt_interval = steps[0].get("interval", config.SCREENSPACE_DEFAULT_INTERVAL)
-            steps[0]["interval"] = (
-                mt_interval * config.SCREENSPACE_FAST_SCAN_INTERVAL_MULTIPLIER
-            )
-        return scan_multitool(
-            video_path,
-            region,
-            steps=steps,
-            start_seconds=params.get("start_seconds", 0.0),
-            end_seconds=params.get("end_seconds"),
-            on_progress=on_progress,
-            cancel_flag=cancel_flag,
-            on_result=on_result,
-            fast_opts=fast_opts,
-        )
-
-
-TOOLS: dict[str, AnalysisTool] = {
-    t.name: t
-    for t in (
-        ColorTool(),
-        ChangeTool(),
-        SimilarityTool(),
-        TextTool(),
-        NumbersTool(),
-        TemplateTool(),
-        FlowTool(),
-        SceneTool(),
-        InactivityTool(),
-        BoundaryTool(),
-        TimelapseTool(),
-        MultitoolTool(),
-    )
-}
-
-
-class ScreenspaceWorker:
-    """Background thread that processes analysis tasks sequentially."""
-
-    def __init__(self) -> None:
-        self._queue: queue.PriorityQueue[Any] = queue.PriorityQueue()
-        self._tasks: dict[str, dict[str, Any]] = {}
-        self._lock = threading.Lock()
-        self._thread: threading.Thread | None = None
-        self._running = False
-        self._paused = threading.Event()
-        self.on_task_complete: Callable[[], None] | None = None
-        self.on_progress_update: Callable[[], None] | None = None
-
-    def start(self) -> None:
-        """Start the worker thread."""
-        self._running = True
-        self._thread = threading.Thread(target=self._run, daemon=True)
-        self._thread.start()
-
-    def restore_tasks(self, tasks: list[dict[str, Any]]) -> None:
-        """Load historical tasks into the worker (completed/failed/cancelled).
-
-        Only non-queued tasks are restored for display; queued tasks are not
-        re-enqueued since the analysis context (video caps) is gone.
-        """
-        with self._lock:
-            for t in tasks:
-                if t.get("id"):
-                    restored = copy.deepcopy(t)
-                    # Tasks now carry ``video_paths`` (multi-video timeline); a
-                    # task persisted before that change has only ``video_path``.
-                    # Normalize on load so resume()/_dispatch never KeyError on it.
-                    if not restored.get("video_paths") and restored.get("video_path"):
-                        restored["video_paths"] = [restored["video_path"]]
-                    self._tasks[t["id"]] = restored
-
-    def stop(self) -> None:
-        """Signal the worker thread to stop."""
-        self._running = False
-        self._queue.put((0, "", _SENTINEL))
-        if self._thread is not None:
-            self._thread.join(timeout=15)
-
-    def enqueue(self, task: dict[str, Any]) -> str:
-        """Add a task to the queue. Returns the task ID."""
-        task_id = task["id"]
-        with self._lock:
-            self._tasks[task_id] = task
-        self._queue.put((task.get("priority", 100), task["created_at"], task_id))
-        return task_id
-
-    def cancel(self, task_id: str) -> bool:
-        """Cancel a queued, running, or paused task. Returns True if cancelled."""
-        with self._lock:
-            task = self._tasks.get(task_id)
-            if task is None:
-                return False
-            if task["status"] in (TASK_STATUS_QUEUED, TASK_STATUS_PAUSED):
-                task["status"] = TASK_STATUS_CANCELLED
-                return True
-            if task["status"] == TASK_STATUS_RUNNING:
-                task["_cancelled"] = True
-                return True
-        return False
-
-    def get_task(self, task_id: str) -> dict[str, Any] | None:
-        """Return a task dict by ID (thread-safe copy)."""
-        with self._lock:
-            task = self._tasks.get(task_id)
-            if task is None:
-                return None
-            return copy.deepcopy(task)
-
-    def get_all_tasks(self) -> list[dict[str, Any]]:
-        """Return all tasks (thread-safe copies).
-
-        Tasks flagged ``_remove_on_finish`` (dismissed while running) are hidden:
-        they are gone as far as the UI and manifest are concerned, even though
-        they linger in ``_tasks`` until their worker thread notices the cancel.
-        """
-        with self._lock:
-            return [
-                copy.deepcopy(t)
-                for t in self._tasks.values()
-                if not t.get("_remove_on_finish")
-            ]
-
-    def reorder(self, task_ids: list[str]) -> bool:
-        """Reorder queued tasks by the given ID sequence.
-
-        Assigns new priorities so that earlier IDs in the list have
-        lower (higher-priority) values.  Drains and re-inserts queue
-        items so the PriorityQueue heap reflects the new ordering.
-        """
-        with self._lock:
-            for i, tid in enumerate(task_ids):
-                task = self._tasks.get(tid)
-                if task and task["status"] == TASK_STATUS_QUEUED:
-                    task["priority"] = i + 1
-
-            # Drain the queue and re-insert with updated priorities
-            pending_items: list[Any] = []
-            while not self._queue.empty():
-                try:
-                    pending_items.append(self._queue.get_nowait())
-                except queue.Empty:
-                    break
-            for priority, created_at, tid in pending_items:
-                task = self._tasks.get(tid) if isinstance(tid, str) else None
-                new_priority = task["priority"] if task else priority
-                self._queue.put((new_priority, created_at, tid))
-        return True
-
-    @property
-    def is_alive(self) -> bool:
-        """Return whether the worker thread is alive."""
-        return self._thread is not None and self._thread.is_alive()
-
-    @property
-    def is_paused(self) -> bool:
-        """Return whether the queue is paused."""
-        return self._paused.is_set()
-
-    def pause(self) -> None:
-        """Pause the queue. Stops the running task so it yields partial results."""
-        self._paused.set()
-        with self._lock:
-            for task in self._tasks.values():
-                if task["status"] == TASK_STATUS_RUNNING:
-                    task["_paused_flag"] = True
-
-    def resume(self) -> None:
-        """Resume the queue. Re-enqueues paused tasks from where they left off."""
-        self._paused.clear()
-        to_resume: list[dict[str, Any]] = []
-        with self._lock:
-            for task in self._tasks.values():
-                if task["status"] == TASK_STATUS_PAUSED:
-                    to_resume.append(task)
-
-        for task in to_resume:
-            progress = task.get("progress", 0.0)
-            params = task.get("parameters", {})
-
-            # Offset multitool chains run the two-phase scan: phase 2 joins
-            # across every frame from the original start, so they cannot resume
-            # mid-stream (advancing start_seconds would drop the pre-pause
-            # frames the join depends on). Restart these from scratch instead —
-            # ``start_seconds`` is left untouched and no partial results carry
-            # over, so the re-run reproduces the full result set.
-            if task.get("type") == "multitool" and _multitool_has_offset(
-                params.get("steps", [])
-            ):
-                with self._lock:
-                    task.pop("_partial_results", None)
-                    task.pop("_progress_offset", None)
-                    task.pop("_progress_scale", None)
-                    task.pop("_paused_flag", None)
-                    task["result"] = []
-                    task["progress"] = 0.0
-                    task["status"] = TASK_STATUS_QUEUED
-                self._queue.put(
-                    (task.get("priority", 100), task["created_at"], task["id"])
-                )
-                continue
-
-            start = params.get("start_seconds", 0.0)
-            end = params.get("end_seconds")
-            if end is None:
-                # Resume math is on the GLOBAL timeline; multi-video end is the
-                # total across all parts.
-                timeline = video.timeline_or_none(task["video_paths"])
-                if timeline is not None:
-                    end = timeline[-1][1] + timeline[-1][2]
-                else:
-                    _, end = _probe_video_meta(task["video_paths"][0])
-            resume_at = start + progress * (end - start)
-
-            with self._lock:
-                task["_partial_results"] = list(task.get("result") or [])
-                task["_progress_offset"] = progress
-                task["_progress_scale"] = max(1.0 - progress, 0.001)
-                task.pop("_paused_flag", None)
-                params["start_seconds"] = resume_at
-                task["status"] = TASK_STATUS_QUEUED
-
-            self._queue.put((task.get("priority", 100), task["created_at"], task["id"]))
-
-    def remove_task(self, task_id: str) -> bool:
-        """Cancel (if active) and fully remove a task."""
-        with self._lock:
-            task = self._tasks.get(task_id)
-            if task is None:
-                return False
-            if task["status"] == TASK_STATUS_RUNNING:
-                # The running scan's cancel_flag looks the task up by id, so the
-                # task must stay in ``_tasks`` for the cancel to land — popping it
-                # here would strand the worker thread, which would run the scan to
-                # completion and keep streaming progress (CPU + SSE spam). Flag it
-                # cancelled + remove-on-finish; ``get_all_tasks`` hides it from the
-                # UI/manifest immediately and ``_execute_task`` pops it once the
-                # scan unwinds.
-                task["_cancelled"] = True
-                task["_remove_on_finish"] = True
-                return True
-            self._tasks.pop(task_id, None)
-            return True
-
-    def drain_new_events(self) -> list[dict[str, Any]]:
-        """Collect and clear ``_generated_events`` from all tasks. Thread-safe."""
-        events: list[dict[str, Any]] = []
-        with self._lock:
-            for t in self._tasks.values():
-                events.extend(t.pop("_generated_events", []))
-        return events
-
-    def _generate_events_from_results(
-        self, task: dict[str, Any], raw_results: list[dict[str, Any]]
-    ) -> list[dict[str, Any]]:
-        return generate_events_from_results(task, raw_results)
-
-    def _generate_heatmap(
-        self, task: dict[str, Any], results: list[dict[str, Any]]
-    ) -> None:
-        """Generate a heatmap PNG for template or flow tasks."""
-        task_type = task.get("type", "")
-        task_id = task["id"]
-        if task_type == "template":
-            heatmap_path = str(
-                Path(utils.get_effective_output_dir()) / f"heatmap_{task_id}.png"
-            )
-            props = video.probe_video_properties(task["video_paths"][0])
-            fw = props.get("width", 1920) if props else 1920
-            fh = props.get("height", 1080) if props else 1080
-            hp = generate_template_heatmap(results, fw, fh, heatmap_path)
-            if hp:
-                task["heatmap"] = Path(hp).name
-            gif_path = str(
-                Path(utils.get_effective_output_dir()) / f"heatmap_{task_id}.gif"
-            )
-            gp = generate_heatmap_gif(
-                results, fw, fh, gif_path, heatmap_type="template"
-            )
-            if gp:
-                task["heatmap_gif"] = Path(gp).name
-        elif task_type == "flow":
-            heatmap_path = str(
-                Path(utils.get_effective_output_dir()) / f"heatmap_{task_id}.png"
-            )
-            rc = task.get("region_coords", {})
-            rw = rc.get("w", 256)
-            rh = rc.get("h", 256)
-            hp = generate_flow_heatmap(results, rw, rh, heatmap_path)
-            if hp:
-                task["heatmap"] = Path(hp).name
-            gif_path = str(
-                Path(utils.get_effective_output_dir()) / f"heatmap_{task_id}.gif"
-            )
-            gp = generate_heatmap_gif(results, rw, rh, gif_path, heatmap_type="flow")
-            if gp:
-                task["heatmap_gif"] = Path(gp).name
-
-    def _run(self) -> None:
-        """Worker loop with concurrent task execution via ThreadPoolExecutor."""
-        max_workers = config.SCREENSPACE_PARALLEL_WORKERS
-        with ThreadPoolExecutor(max_workers=max_workers) as pool:
-            active: dict[str, Future[None]] = {}
-
-            while self._running:
-                try:
-                    # 1. Collect completed futures
-                    done_ids = [tid for tid, f in active.items() if f.done()]
-                    for tid in done_ids:
-                        future = active.pop(tid)
-                        try:
-                            future.result()
-                        except Exception as exc:
-                            utils.warning_print(f"Worker task {tid} raised: {exc}")
-                        if self.on_task_complete:
-                            try:
-                                self.on_task_complete()
-                            except Exception as exc:
-                                utils.warning_print(
-                                    f"on_task_complete callback failed: {exc}"
-                                )
-                        if self.on_progress_update:
-                            try:
-                                self.on_progress_update()
-                            except Exception as exc:
-                                utils.warning_print(
-                                    f"on_progress_update callback failed: {exc}"
-                                )
-
-                    # 2. If paused, wait
-                    if self._paused.is_set():
-                        time.sleep(0.25)
-                        continue
-
-                    # 3. Submit new tasks if capacity available
-                    if len(active) >= max_workers:
-                        time.sleep(0.1)
-                        continue
-
-                    try:
-                        item = self._queue.get(timeout=0.25)
-                    except queue.Empty:
-                        continue
-
-                    priority, created_at, task_id = item
-                    if task_id is _SENTINEL:
-                        # Wait for active tasks to finish
-                        for drain_tid, f in active.items():
-                            try:
-                                f.result(timeout=10)
-                            except Exception as exc:
-                                utils.debug_print(
-                                    f"Task {drain_tid} raised during shutdown: {exc}"
-                                )
-                        if self.on_task_complete:
-                            try:
-                                self.on_task_complete()
-                            except Exception as exc:
-                                utils.warning_print(
-                                    f"on_task_complete callback failed during shutdown: {exc}"
-                                )
-                        break
-
-                    if self._paused.is_set():
-                        self._queue.put(item)
-                        time.sleep(0.25)
-                        continue
-
-                    with self._lock:
-                        task = self._tasks.get(task_id)
-                        if task is None or task["status"] != TASK_STATUS_QUEUED:
-                            continue
-                        task["status"] = TASK_STATUS_RUNNING
-
-                    future = pool.submit(self._execute_task, task)
-                    active[task_id] = future
-
-                except Exception as exc:
-                    utils.warning_print(f"Worker loop error: {exc}")
-
-    def _execute_task(self, task: dict[str, Any]) -> None:
-        """Dispatch task to the appropriate workflow function."""
-        task_id = task["id"]
-
-        # Seed incremental results (includes partial results on resume)
-        with self._lock:
-            t = self._tasks.get(task_id)
-            if t:
-                t["result"] = list(t.get("_partial_results", []))
-                t["_raw_results"] = list(t.get("_partial_results", []))
-
-        # Per-task throttle for SSE notifications (avoids cross-task race)
-        last_notify: list[float] = [0.0]
-
-        def _throttled_notify() -> None:
-            now = time.monotonic()
-            if now - last_notify[0] >= 0.5:
-                last_notify[0] = now
-                if self.on_progress_update:
-                    self.on_progress_update()
-
-        def _on_progress(progress: float) -> None:
-            with self._lock:
-                t = self._tasks.get(task_id)
-                if t and not t.get("_paused_flag"):
-                    offset = t.get("_progress_offset", 0.0)
-                    scale = t.get("_progress_scale", 1.0)
-                    t["progress"] = min(offset + progress * scale, 1.0)
-            _throttled_notify()
-
-        def _cancel_flag() -> bool:
-            with self._lock:
-                t = self._tasks.get(task_id)
-                return bool(t and (t.get("_cancelled") or t.get("_paused_flag")))
-
-        def _on_result(result_dict: dict[str, Any]) -> None:
-            with self._lock:
-                t = self._tasks.get(task_id)
-                if t:
-                    if isinstance(t.get("result"), list):
-                        t["result"].append(result_dict)
-                    if isinstance(t.get("_raw_results"), list):
-                        t["_raw_results"].append(result_dict)
-                    if t.get("parameters", {}).get("detect_first"):
-                        t["_cancelled"] = True
-            _throttled_notify()
-
-        try:
-            result = self._dispatch(task, _on_progress, _cancel_flag, _on_result)
-            with self._lock:
-                t = self._tasks.get(task_id)
-                if t:
-                    if t.get("_paused_flag"):
-                        t["status"] = TASK_STATUS_PAUSED
-                        t["result"] = result
-                    elif t.get("_cancelled") and not t.get("parameters", {}).get(
-                        "detect_first"
-                    ):
-                        t["status"] = TASK_STATUS_CANCELLED
-                    else:
-                        # Normal completion or detect_first early stop
-                        partial = t.pop("_partial_results", None)
-                        if partial and isinstance(result, list):
-                            result = partial + result
-                        t["status"] = TASK_STATUS_COMPLETED
-                        t["result"] = result
-                        t["progress"] = 1.0
-                        t.pop("_progress_offset", None)
-                        t.pop("_progress_scale", None)
-                        raw = t.pop("_raw_results", [])
-                        t["_generated_events"] = self._generate_events_from_results(
-                            t, raw
-                        )
-                        if isinstance(result, list) and result:
-                            self._generate_heatmap(t, result)
-                    t["completed_at"] = datetime.now(timezone.utc).isoformat()
-        except Exception as exc:
-            with self._lock:
-                t = self._tasks.get(task_id)
-                if t:
-                    t["status"] = TASK_STATUS_FAILED
-                    t["error"] = str(exc)
-                    t["completed_at"] = datetime.now(timezone.utc).isoformat()
-        finally:
-            # A task dismissed while running was kept in _tasks so its cancel
-            # could land; now that the scan has unwound, drop it for good.
-            with self._lock:
-                t = self._tasks.get(task_id)
-                if t is not None and t.get("_remove_on_finish"):
-                    self._tasks.pop(task_id, None)
-
-    def _dispatch(
-        self,
-        task: dict[str, Any],
-        on_progress: Callable[[float], None],
-        cancel_flag: Callable[[], bool],
-        on_result: Callable[[dict[str, Any]], None] | None = None,
-    ) -> Any:
-        """Route task to the correct tool via the ``TOOLS`` registry.
-
-        Builds the shared fast-scan ``fast_opts`` payload here so individual
-        tool classes only need to declare ``fast_scan_region_dim`` /
-        ``supports_fast_scan`` / ``fast_scan_extra_opts`` as class attributes.
-        """
-        task_type = task["type"]
-        tool = TOOLS.get(task_type)
-        if tool is None:
-            raise ValueError(f"Unknown task type: {task_type}")
-
-        # Shallow copy so fast-scan interval scaling is not persisted on the
-        # task dict (pause/resume re-dispatches the same parameters).
-        params = dict(task.get("parameters", {}))
-        scan_mode = params.get("scan_mode", "normal")
-        fast_opts: dict[str, Any] | None = None
-        if scan_mode == "fast" and tool.supports_fast_scan:
-            multiplier = config.SCREENSPACE_FAST_SCAN_INTERVAL_MULTIPLIER
-            if params.get("interval", 0) > 0:
-                params["interval"] = params["interval"] * multiplier
-            fast_opts = {
-                "phash_skip": True,
-                "max_region_dim": tool.fast_scan_region_dim,
-                **tool.fast_scan_extra_opts,
-            }
-
-        video_paths = task["video_paths"]
-        timeline = video.timeline_or_none(video_paths)
-        if timeline is None:
-            return tool.scan(
-                video_paths[0],
-                task["region_coords"],
-                params,
-                task_id=task["id"],
-                scan_mode=scan_mode,
-                on_progress=on_progress,
-                cancel_flag=cancel_flag,
-                on_result=on_result,
-                fast_opts=fast_opts,
-            )
-
-        # Multi-video: map the task's GLOBAL [start, end] range onto the timeline
-        # and scan each spanned sub-video at its local offsets. Emitted result
-        # times are shifted back to the global timeline and tagged with the
-        # sub-video they came from so events line up with clips/transcripts.
-        total = timeline[-1][1] + timeline[-1][2]
-        global_start = params.get("start_seconds", 0.0) or 0.0
-        global_end = params.get("end_seconds")
-        if global_end is None:
-            global_end = total
-        pieces = utils.map_global_range_to_segments(timeline, global_start, global_end)
-        if not pieces:
-            return []
-        piece_durations = [le - ls for _index, ls, le in pieces]
-        span = sum(piece_durations) or 1.0
-        accumulated = 0.0
-        all_results: list[dict[str, Any]] = []
-        for (index, local_start, local_end), piece_dur in zip(pieces, piece_durations):
-            if cancel_flag():
-                break
-            cumulative = timeline[index][2]
-            source_name = Path(timeline[index][0]).name
-            frac_start = accumulated / span
-            frac_end = (accumulated + piece_dur) / span
-
-            def piece_progress(
-                p: float, _a: float = frac_start, _b: float = frac_end
-            ) -> None:
-                on_progress(_a + p * (_b - _a))
-
-            def piece_on_result(
-                rd: dict[str, Any],
-                _cum: int = cumulative,
-                _src: str = source_name,
-            ) -> None:
-                if on_result is not None:
-                    shifted = dict(rd)
-                    _offset_result_times(shifted, _cum)
-                    shifted["_source_video"] = _src
-                    on_result(shifted)
-
-            piece_params = {
-                **params,
-                "start_seconds": local_start,
-                "end_seconds": local_end,
-            }
-            piece_results = tool.scan(
-                timeline[index][0],
-                task["region_coords"],
-                piece_params,
-                task_id=task["id"],
-                scan_mode=scan_mode,
-                on_progress=piece_progress,
-                cancel_flag=cancel_flag,
-                on_result=piece_on_result,
-                fast_opts=fast_opts,
-            )
-            for rd in piece_results or []:
-                _offset_result_times(rd, cumulative)
-                rd["_source_video"] = source_name
-                all_results.append(rd)
-            accumulated += piece_dur
-        on_progress(1.0)
-        return all_results
-
-
-# ---------------------------------------------------------------------------
-# Manifest I/O
-# ---------------------------------------------------------------------------
-
-
-def _empty_screenspace_manifest() -> dict[str, Any]:
-    return {
-        "regions": {},
-        "tasks": [],
-        "events": [],
-        "stashes": [],
-        "per_participant": {},
-        "pins": {},
-    }
-
-
-def load_screenspace_manifest() -> dict[str, Any]:
-    """Load the screenspace manifest from the output directory."""
-    return utils.load_json_manifest(
-        config.SCREENSPACE_MANIFEST_FILENAME, default=_empty_screenspace_manifest()
-    )
-
-
-def save_screenspace_manifest(
-    regions: dict[str, dict[str, Any]],
-    tasks: list[dict[str, Any]],
-    events: list[dict[str, Any]] | None = None,
-    stashes: list[dict[str, Any]] | None = None,
-    per_participant: dict[str, dict[str, Any]] | None = None,
-    pins: dict[str, list[dict[str, Any]]] | None = None,
-) -> Path | None:
-    """Write the screenspace manifest to disk.
-
-    Strips internal fields (prefixed with ``_``) from tasks before writing.
-    Returns the manifest path on success, or ``None`` on failure.
-    """
-    clean_tasks = []
-    for task in tasks:
-        ct = {k: v for k, v in task.items() if not k.startswith("_")}
-        if "parameters" in ct:
-            _binary_keys = (
-                "reference_frame",
-                "template_image",
-                "template_mask",
-                "reference_scenes",
-            )
-            ct["parameters"] = {
-                k: v for k, v in ct["parameters"].items() if k not in _binary_keys
-            }
-            # Strip binary data and internal coords from multitool step parameters
-            if "steps" in ct["parameters"]:
-                _step_strip_keys = _binary_keys + ("region_coords",)
-                ct["parameters"]["steps"] = [
-                    {k: v for k, v in s.items() if k not in _step_strip_keys}
-                    for s in ct["parameters"]["steps"]
-                ]
-        # Strip flow_grid from results (large per-frame data, not needed on disk)
-        if isinstance(ct.get("result"), list):
-            ct["result"] = [
-                {k: v for k, v in r.items() if k != "flow_grid"} for r in ct["result"]
-            ]
-        clean_tasks.append(ct)
-    if pins is None:
-        existing = load_screenspace_manifest()
-        existing_pins = existing.get("pins")
-        pins_payload = existing_pins if isinstance(existing_pins, dict) else {}
-    else:
-        pins_payload = pins
-
-    payload = utils.sanitize_floats(
-        {
-            "regions": regions,
-            "tasks": clean_tasks,
-            "events": events or [],
-            "stashes": stashes or [],
-            "per_participant": per_participant or {},
-            "pins": pins_payload,
-        }
-    )
-    return utils.save_json_manifest(
-        config.SCREENSPACE_MANIFEST_FILENAME,
-        payload,
-        warn_label="screenspace manifest",
-    )
-
-
-def _offset_result_times(result: dict[str, Any], offset: int) -> None:
-    """Shift a scan result's time fields by *offset* seconds (in place).
-
-    Maps a sub-video's local result times back onto the participant's global
-    timeline for multi-video scans. Covers point events (``timestamp``) and span
-    events (``start``/``end``, e.g. inactivity). A no-op when offset is 0.
-    """
-    if not offset:
-        return
-    for key in ("timestamp", "start", "end"):
-        value = result.get(key)
-        if isinstance(value, (int, float)):
-            result[key] = value + offset
-
-
-def generate_events_from_results(
-    task: dict[str, Any], raw_results: list[dict[str, Any]]
-) -> list[dict[str, Any]]:
-    """Convert raw task results into ScreenspaceEvent records."""
-    task_type = task.get("type", "")
-    if task_type == "timelapse":
-        return []
-    events: list[dict[str, Any]] = []
-    for r in raw_results:
-        ts = r.get("timestamp", r.get("start", 0.0))
-        confidence = _extract_confidence(task_type, r)
-        metadata: dict[str, Any] = {}
-        if task_type == "change":
-            metadata["magnitude"] = r.get("magnitude", 0.0)
-        elif task_type == "similarity":
-            metadata["score"] = r.get("score", 0.0)
-        elif task_type == "text":
-            metadata["text_found"] = r.get("text_found", "")
-        elif task_type == "numbers":
-            metadata["value"] = r.get("number_found", 0)
-        elif task_type == "template":
-            metadata["match_count"] = r.get("match_count", 0)
-            metadata["best_score"] = r.get("best_score", 0.0)
-        elif task_type == "flow":
-            metadata["magnitude"] = r.get("magnitude", 0.0)
-            metadata["angle"] = r.get("angle", 0.0)
-        elif task_type == "scene":
-            metadata["scene_name"] = r.get("scene_name", "")
-            metadata["score"] = r.get("score", 0.0)
-        elif task_type == "multitool":
-            metadata["tool_types"] = r.get("tool_types", [])
-            metadata["steps"] = r.get("steps", [])
-        elif task_type == "inactivity":
-            metadata["duration"] = r.get("duration", 0.0)
-            metadata["avg_distance"] = r.get("avg_distance", 0.0)
-        elif task_type == "boundary":
-            metadata["distance"] = r.get("distance", 0.0)
-        ev = create_event(task, ts, confidence, metadata)
-        # Multi-video scans tag each result with the sub-video it came from.
-        source_override = r.get("_source_video")
-        if source_override:
-            ev["source_video"] = source_override
-        if task_type == "inactivity" and "end" in r:
-            ev["time_out"] = round(r["end"], 2)
-        if task_type == "boundary":
-            # Boundaries are for orientation, not clip candidacy. The
-            # navigational flag lets Studio intake hide them by default.
-            ev["navigational"] = True
-        events.append(ev)
-    return events
-
-
-def create_event(
-    task: dict[str, Any],
-    timestamp: float,
-    confidence: float,
-    metadata: dict[str, Any] | None = None,
-) -> dict[str, Any]:
-    """Create a ScreenspaceEvent from a task result entry."""
-    event_label = task.get("parameters", {}).get("event_label", "")
-    if not event_label:
-        event_label = task["type"] + ": " + task.get("region", "")
-    return {
-        "id": f"ev_{uuid.uuid4().hex[:8]}",
-        "source_video": task.get("source_video", ""),
-        "participant": task.get("participant", ""),
-        "detector": task["type"],
-        "event_type": event_label,
-        "time_in": round(timestamp, 2),
-        "time_out": round(timestamp, 2),
-        "confidence": round(max(0.0, min(1.0, confidence)), 4),
-        "metadata": metadata or {},
-        "excluded": False,
-        "task_id": task["id"],
-        "region": task.get("region", ""),
-    }
+from screenspace_ocr import (
+    _get_ocr_reader,
+    _ocr_readers,
+    _preprocess_for_ocr,
+    _score_numbers_readings,
+    _score_text_readings,
+    run_calibration_ocr,
+)
+from screenspace_frames import (
+    _ffmpeg_pipe_frames,
+    _scan_via_ffmpeg_pipe,
+    build_timelapse_command,
+    scan_video_frames,
+    scan_video_full_frames,
+)
+from screenspace_scans import (
+    generate_timelapse,
+    scan_boundaries,
+    scan_changes,
+    scan_color,
+    scan_flow,
+    scan_inactivity,
+    scan_numbers,
+    scan_scene,
+    scan_similarity,
+    scan_template,
+    scan_text,
+)
+from screenspace_heatmap import (
+    generate_flow_heatmap,
+    generate_heatmap_gif,
+    generate_template_heatmap,
+)
+from screenspace_tools import (
+    TOOLS,
+    AnalysisTool,
+    BoundaryTool,
+    ChangeTool,
+    ColorTool,
+    FlowTool,
+    InactivityTool,
+    MultitoolTool,
+    NumbersTool,
+    SceneTool,
+    SimilarityTool,
+    TemplateTool,
+    TextTool,
+    TimelapseTool,
+    _extract_confidence,
+    check_frame_for_tool,
+    score_frame_for_tool,
+)
+from screenspace_multitool import scan_multitool, score_multitool_frame
+from screenspace_manifest import (
+    TASK_STATUS_CANCELLED,
+    TASK_STATUS_COMPLETED,
+    TASK_STATUS_FAILED,
+    TASK_STATUS_PAUSED,
+    TASK_STATUS_QUEUED,
+    TASK_STATUS_RUNNING,
+    _empty_screenspace_manifest,
+    _offset_result_times,
+    create_event,
+    create_task,
+    generate_events_from_results,
+    load_screenspace_manifest,
+    save_screenspace_manifest,
+)
+from screenspace_worker import ScreenspaceWorker
